% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fittingFunctions.R
\docType{methods}
\name{a4a}
\alias{a4a}
\title{a4a}
\usage{
a4a(...)
}
\arguments{
\item{...}{Additional argument list that might not ever be used.}
}
\description{
This was the old stock assessment framework user's interface. It was replaced by \code{sca} or the advanced version \code{a4aSCA}.
}

