% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a4aFitCatchDiagn-class.R
\docType{class}
\name{a4aFitCatchDiagn-class}
\alias{a4aFitCatchDiagn-class}
\alias{computeCatchDiagnostics}
\alias{a4aFitCatchDiagn}
\alias{a4aFitCatchDiagn-methods}
\title{S4 class \code{a4aFitCatchDiagn}}
\usage{
computeCatchDiagnostics(object, ...)
}
\arguments{
\item{object}{object of relevant class (see signature of method)}

\item{...}{additional argument list that might never be used}

\item{stock}{\code{FLStock} object used to fit the model}

\item{indices}{\code{FLIndices} object used to fit the model}
}
\description{
The \code{a4aFitCatchDiagn} class extends \code{FLQuants} to store information to run diagnostics on aggregated catch estimated  by the a4a stock assessment fit.
}
\examples{
data(ple4)
data(ple4.index)
obj <- sca(stock=ple4, indices=FLIndices(ple4.index))
flqs <- residuals(obj, ple4, FLIndices(idx=ple4.index))
}
