% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a4aFitMCMC-class.R
\docType{class}
\name{a4aFitMCMC-class}
\alias{a4aFitMCMC-class}
\alias{a4aFitMCMC}
\alias{a4aFitMCMC-methods}
\alias{a4aFitSA,a4aFitMCMC-method}
\alias{a4aFit,a4aFitMCMC-method}
\alias{as.mcmc}
\alias{as.mcmc-methods}
\alias{as.mcmc,a4aFitMCMC-method}
\alias{burnin}
\alias{burnin-methods}
\alias{burnin,a4aFitMCMC-method}
\title{S4 class \code{a4aFitMCMC}}
\usage{
a4aFitMCMC(...)

a4aFitMCMC(...)

\S4method{a4aFitSA}{a4aFitMCMC}(object, ...)

\S4method{a4aFit}{a4aFitMCMC}(object, ...)

as.mcmc(x, ...)

\S4method{as.mcmc}{a4aFitMCMC}(x, ...)

burnin(object, ...)

\S4method{burnin}{a4aFitMCMC}(object, burnin)
}
\arguments{
\item{...}{additional argument list that might never be used}

\item{object}{object of relevant class (see signature of method)}

\item{x}{an object to be coerced into mcmc}

\item{burnin}{a numeric with the number of iterations to be removed}

\item{object}{a \code{a4aFitMCMC} object}
}
\description{
The \code{a4aFitMCMC} class extends \code{a4aFitSA} to store information about the MCMC run.
}
\section{Slots}{

\describe{
   \item{name}{A character vector for the object name.}
   \item{desc}{A textual description of the object contents.}
   \item{range}{A named numeric vector with various values of quant and year ranges, plusgroup, fishing mortality ranges, etc.}
   \item{call}{The function call}
   \item{clock}{Information on call duration}
   \item{fitSumm}{Fit summary}
   \item{stock.n}{Estimates of stock numbers-at-age}
   \item{harvest}{Estimates of fishing mortality at age}
   \item{catch.n}{Estimates of catch numbers-at-age}
   \item{index}{Estimates of survey or CPUE indices-at-age}
   \item{mcmc}{An object of class \code{SCAMCMC} with information about the MCMC run}
}
}

\section{Accessors}{

All slots in the class have accessor and replacement methods defined that
allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot.
A numeric vector can also be used when replacing FLQuant slots, and the
vector will be used to substitute the values in the slot, but not its other
attributes.
}

\section{Constructor}{

A construction method exists for this class that can take named arguments for
any of its slots. All slots are then created to match the requirements of the
class validity. If an unnamed \code{FLQuant} object is provided, this is used
for sizing, but not for populating any slot.
}

\examples{
data(ple4)
data(ple4.index)

obj <- sca(stock=ple4, indices=FLIndices(ple4.index), fit="assessment")
obj

slotNames(obj)
clock(obj)
fitSumm(obj)

flq <- stock.n(obj)
is(flq)
flq <- index(obj)
is(flq)

logLik(obj)
AIC(obj)
BIC(obj)

is(pars(obj))
}
