\name{a4aFitResiduals-class}
\alias{a4aFitResiduals-class}
\title{a4aFitResiduals extends \code{"FLQuant"} class.}
\description{
  Some details about this class and my plans for it in the
  body.
}
\details{
  \describe{ \item{myslot1}{A logical keeping track of
  something.}

  \item{myslot2}{An integer specifying something else.}

  \item{myslot3}{A data.frame holding some data.} }
}

