\docType{methods}
\name{SCAPars}
\alias{a4aFitSA}
\alias{a4aFitSA,a4aFitSA-method}
\alias{fCovar}
\alias{fFrml}
\alias{fmodel}
\alias{fPars}
\alias{n1model}
\alias{qCovar}
\alias{qFrml}
\alias{qmodel}
\alias{qPars}
\alias{SCAPars}
\alias{srCovar}
\alias{srFrml}
\alias{srmodel}
\alias{srPars}
\alias{stkmodel}
\alias{vCovar}
\alias{vFrml}
\alias{vmodel}
\alias{vPars}
\title{Calculate the median accross iterations}
\arguments{
  \item{object}{an FLQuant with iters}

  \item{...}{Additional argument list that might not ever
  be used.}

  \item{object}{an FLQuant with iters}

  \item{...}{Additional argument list that might not ever
  be used.}

  \item{object}{an FLQuant with iters}

  \item{...}{Additional argument list that might not ever
  be used.}

  \item{object}{an FLQuant with iters}

  \item{...}{Additional argument list that might not ever
  be used.}

  \item{object}{an FLQuant with iters}

  \item{...}{Additional argument list that might not ever
  be used.}

  \item{object}{an FLQuant with iters}

  \item{...}{Additional argument list that might not ever
  be used.}

  \item{object}{an FLQuant with iters}

  \item{...}{Additional argument list that might not ever
  be used.}

  \item{object}{an FLQuant with iters}

  \item{...}{Additional argument list that might not ever
  be used.}

  \item{object}{an FLQuant with iters}

  \item{...}{Additional argument list that might not ever
  be used.}

  \item{object}{an FLQuant with iters}

  \item{...}{Additional argument list that might not ever
  be used.}

  \item{object}{an FLQuant with iters}

  \item{...}{Additional argument list that might not ever
  be used.}

  \item{object}{an FLQuant with iters}

  \item{...}{Additional argument list that might not ever
  be used.}

  \item{object}{an FLQuant with iters}

  \item{...}{Additional argument list that might not ever
  be used.}

  \item{object}{an FLQuant with iters}

  \item{...}{Additional argument list that might not ever
  be used.}

  \item{object}{an FLQuant with iters}

  \item{...}{Additional argument list that might not ever
  be used.}

  \item{object}{an FLQuant with iters}

  \item{...}{Additional argument list that might not ever
  be used.}

  \item{object}{an FLQuant with iters}

  \item{...}{Additional argument list that might not ever
  be used.}

  \item{object}{an FLQuant with iters}

  \item{...}{Additional argument list that might not ever
  be used.}

  \item{object}{an FLQuant with iters}

  \item{...}{Additional argument list that might not ever
  be used.}
}
\value{
  an FLQuant

  an FLQuant

  an FLQuant

  an FLQuant

  an FLQuant

  an FLQuant

  an FLQuant

  an FLQuant

  an FLQuant

  an FLQuant

  an FLQuant

  an FLQuant

  an FLQuant

  an FLQuant

  an FLQuant

  an FLQuant

  an FLQuant

  an FLQuant

  an FLQuant
}
\description{
  Calculate the median accross iterations

  Calculate the median accross iterations

  Title

  Calculate the median accross iterations

  Title

  Calculate the median accross iterations

  Title

  Calculate the median accross iterations

  Title

  Calculate the median accross iterations

  Title

  Calculate the median accross iterations

  Title

  Calculate the median accross iterations

  Title

  Calculate the median accross iterations

  Title

  Calculate the median accross iterations

  Title

  Calculate the median accross iterations

  Title

  Calculate the median accross iterations

  Title

  Calculate the median accross iterations

  Title

  Calculate the median accross iterations

  Title

  Calculate the median accross iterations

  Title

  Calculate the median accross iterations

  Title

  Calculate the median accross iterations

  Title

  Calculate the median accross iterations

  Title

  Calculate the median accross iterations

  Title
}
\examples{
data(ple4)
data(ple4)
data(ple4)
data(ple4)
data(ple4)
data(ple4)
data(ple4)
data(ple4)
data(ple4)
data(ple4)
data(ple4)
data(ple4)
data(ple4)
data(ple4)
data(ple4)
data(ple4)
data(ple4)
data(ple4)
data(ple4)
}
\seealso{
  \code{\link{print}} and \code{\link{cat}}

  \code{\link{print}} and \code{\link{cat}}

  \code{\link{print}} and \code{\link{cat}}

  \code{\link{print}} and \code{\link{cat}}

  \code{\link{print}} and \code{\link{cat}}

  \code{\link{print}} and \code{\link{cat}}

  \code{\link{print}} and \code{\link{cat}}

  \code{\link{print}} and \code{\link{cat}}

  \code{\link{print}} and \code{\link{cat}}

  \code{\link{print}} and \code{\link{cat}}

  \code{\link{print}} and \code{\link{cat}}

  \code{\link{print}} and \code{\link{cat}}

  \code{\link{print}} and \code{\link{cat}}

  \code{\link{print}} and \code{\link{cat}}

  \code{\link{print}} and \code{\link{cat}}

  \code{\link{print}} and \code{\link{cat}}

  \code{\link{print}} and \code{\link{cat}}

  \code{\link{print}} and \code{\link{cat}}

  \code{\link{print}} and \code{\link{cat}}
}

