% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/a4aM-class.R
\docType{class}
\name{a4aM}
\alias{a4aM}
\alias{a4aM,missing-method}
\alias{a4aM-class}
\alias{a4aM-methods}
\alias{level}
\alias{level,a4aM-method}
\alias{level-methods}
\alias{level<-}
\alias{level<-,a4aM-method}
\alias{level<--methods}
\alias{shape}
\alias{shape,a4aM-method}
\alias{shape-methods}
\alias{shape<-}
\alias{shape<-,a4aM-method}
\alias{shape<--methods}
\alias{show,a4aM-method}
\alias{trend}
\alias{trend,a4aM-method}
\alias{trend-methods}
\alias{trend<-}
\alias{trend<-,a4aM-method}
\alias{trend<--methods}
\title{Natural mortality class}
\usage{
a4aM(object, ...)

\S4method{show}{a4aM}(object)

shape(object, ...)

shape(object) <- value

level(object, ...)

level(object) <- value

trend(object, ...)

trend(object) <- value
}
\arguments{
\item{object}{todo}

\item{...}{Additional argument list that might not ever be used.}

\item{value}{todo}
}
\description{
Class definition (slots), constructors, accessors, replacement (when relevant) and common methods.
}
\section{Slot}{

\describe{
	\item{\code{shape}}{the shape of M by age}
	\item{\code{level}}{the mean level of M in a range of ages, which will be used to scale the \code{shape}}
	\item{\code{trend}}{the yearly trend in M}
}
}

\section{Accessors}{

All slots in the class have accessor and replacement methods defined that
allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot.
A numeric vector can also be used when replacing FLQuant slots, and the
vector will be used to substitute the values in the slot, but not its other
attributes.
}

\section{Constructor}{

A construction method exists for this class that can take named arguments for
any of its slots. All slots are then created to match the requirements of the
class validity. If an unnamed \code{FLQuant} object is provided, this is used
for sizing but not stored in any slot.
}
\examples{
mod1 <- FLModelSim(model=~exp(-age-0.5))
mod2 <- FLModelSim(model=~1.5*k, params=FLPar(k=0.4))
m1 <- a4aM(shape=mod1, level=mod2)
}

