\name{bp}
\alias{bp}
\title{adds bubbles to a plot - haha!}
\usage{
  bp(x, y, v, scale = 3, ...)
}
\arguments{
  \item{x}{the x location of the bubbles}

  \item{y}{the y location of the bubbles}

  \item{v}{a vector of bubble sizes}

  \item{scale}{a scaling factor}

  \item{...}{extra arguments to plot and points}
}
\value{
  a pointer to the environment in which summaries of the
  data reside
}
\description{
  adds bubbles to a plot - haha!
}
\note{
  \code{extractData} is intended to be used internally
}
\author{
  anders nielson \email{an@dtu.aqua.dk}
}

