\name{calcLogLik}
\alias{calcLogLik}
\title{Method to compute log likelihood}
\usage{
  calcLogLik(b, object, stock)
}
\arguments{
  \item{b}{a \code{a4aFitSA} object}

  \item{object}{a \code{a4a??} object}

  \item{stock}{a \code{FLStock} object}
}
\value{
  ??
}
\description{
  Method to compute log likelihood
}

