\docType{methods}
\name{catch.lhat}
\alias{catch.lhat}
\alias{catch.lhat,FLa4aFit-method}
\title{Calculate the median accross iterations}
\arguments{
  \item{object}{an FLQuant with iters}

  \item{...}{Additional argument list that might not ever
  be used.}
}
\value{
  an FLQuant
}
\description{
  Calculate the median accross iterations

  Hello
}
\seealso{
  \code{\link{print}} and \code{\link{cat}}
}

