% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoother_constructors.R
\name{constructor for smoother}
\alias{constructor for smoother}
\alias{smooth.construct.cgmrf.smooth.spec}
\alias{Predict.matrix.cgmrf.smooth}
\title{Reduced Rank Constrained Random Walk Smooths}
\usage{
\method{smooth.construct}{cgmrf.smooth.spec}(object, data, knots)

\method{Predict.matrix}{cgmrf.smooth}(object, data)
}
\arguments{
\item{object}{For the smooth.construct method a smooth specification object,
usually generated by a term
s(x,..., bs="cgmrf", xt = list(gaps, weight, xrange)).
x is a factor variable giving labels for distinct years or ages,
and the xt argument is obligatory: see details. For the
Predict.Matrix method an object of class "cgmrf.smooth"
produced by the smooth.construct method.}

\item{data}{a list containing just the data (including any by variable)
required by this term, with names corresponding to object$term.}

\item{knots}{not used.}
}
\value{
An object of class "cgmrf.smooth" or a matrix mapping the
       coefficients of the cgmrf smooth to the predictions for the years/ages
       listed in data.
}
\description{
Function to construct a smoother object for use in mgcv
}
\details{
some details on the xt list required
}
\examples{
# simulate a sine trend with a little noise
dat <- data.frame(year = 1980:2020)
dat$trend <- 10 * sin(dat$year * 2 * pi / 40)
dat$y <- dat$trend + rnorm(nrow(dat), 0, 0.5)
gaps <- c(1970:1985, 1997:2004)
dat$gap <- dat$year \%in\% gaps
plot(y ~ year, data = dat)
points(y ~ year, data = dat[!dat$gap,], pch = 16)
lines(trend ~ year, data = dat, col = "blue", lty = 2)
form <- y ~ s(year, k = 9, bs = "cgmrf", m = 2, fx = TRUE,
                  xt = list(gaps = gaps, weight = 0.01, xrange = c(1980, 2020)))

library(mgcv)
 g <- gam(formula(form), data = dat[!dat$gap,])
 pred <- predict(g, newdata = dat, se.fit = TRUE)
 lines(dat$year, pred$fit)
 lines(dat$year, pred$fit + pred$se.fit*2, lty = 2)
 lines(dat$year, pred$fit - pred$se.fit*2, lty = 2)
}
