% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{defaultSubmodels}
\alias{defaultSubmodels}
\alias{defaultFmod}
\alias{defaultQmod}
\alias{defaultN1mod}
\alias{defaultVmod}
\alias{defaultSRmod}
\title{Default submodels}
\usage{
defaultFmod(stock, dfm = c(0.5, 0.7))

defaultQmod(indices, dfm = 0.6)

defaultN1mod(stock)

defaultVmod(stock, indices)

defaultSRmod(stock)
}
\arguments{
\item{stock}{an \code{FLStock}.}

\item{dfm}{fraction of data points used to set the number of knots.}

\item{indices}{an \code{FLIndices}.}
}
\value{
a formula, or a list of formulas.
}
\description{
Formulas used by \code{\link[=sca]{sca()}} when a submodel is not given. The number of
knots scales with the number of ages and years in the data.
}
