% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fittingFunctions.R
\docType{methods}
\name{deprecated}
\alias{deprecated}
\alias{a4aSCA}
\title{deprecated}
\usage{
a4aSCA(...)
}
\arguments{
\item{...}{additional argument list that might never be used}
}
\description{
Deprecated methods.
}
