\name{Dfunc}
\alias{Dfunc}
\title{creates a matrix operator for the nth difference}
\usage{
  Dfunc(n)
}
\arguments{
  \item{n}{the length of the vector for difference are to
  be calculated}
}
\value{
  a matrix which functions as the difference operator
}
\description{
  creates a matrix operator for the nth difference
}
\author{
  Colin Millar \email{colin.millar@jrc.ec.europa.eu}
}

