\name{distr<-}
\alias{distr<-}
\title{set distribution}
\arguments{
  \item{object}{a \code{a4aGr} object}

  \item{value}{a \code{character} object}
}
\value{
  a \code{a4aGr} object
}
\description{
  set method for \code{a4aGr} object's slot \code{distr}.
}
\author{
  EJ \email{ernesto.jardim@jrc.ec.europa.eu}
}

