% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a4amse-is.R
\name{eff.is}
\alias{eff.is}
\title{effort implementation function}
\usage{
eff.is(stk, ctrl, ay, tracking)
}
\arguments{
\item{stk}{The perceived FLStock.}

\item{ay}{The year for which the target F is set, based on the SSB in year (ay - control$ssb_lag).}

\item{imp_control}{A list with the elements: nsqy, delta_tac_min, delta_tac_max}
}
\description{
effort implementation function
}
