\docType{package}
\name{FLa4a-package}
\alias{FLa4a}
\alias{FLa4a-package}
\title{A flexible age based stock assessment model}
\description{
  A simple and robust statistical catch at age model
}
\details{
  \tabular{ll}{ Package: \tab FLa4a \cr Type: \tab Package
  \cr Version: \tab 1.0 \cr Date: \tab 2013-06-01 \cr
  License: \tab GPL (>= 2) \cr LazyLoad: \tab yes \cr }
}
\examples{
data(ple4)
data(ple4.index)
}
\author{
  Colin Millar \email{colin.millar@jrc.ec.europa.eu}

  Ernesto Jardim \email{ernesto.jardim@jrc.ec.europa.eu}
}
\references{
  none-yet
}
\seealso{
  \code{\link{paste}}
}
\keyword{package}

