% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a4amse-hcr.R
\name{fxdFtrg.hcr}
\alias{fxdFtrg.hcr}
\title{A fixed target f}
\usage{
fxdFtrg.hcr(stk, ftrg, ay, tracking)
}
\arguments{
\item{stk}{The perceived FLStock.}

\item{control}{A list with the element ftrg (numeric).}
}
\description{
No matter what get F= Ftarget
The control argument is a list of parameters used by the HCR.
}
