% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-methods.R
\docType{methods}
\name{genFLStock}
\alias{genFLStock}
\alias{genFLStock-methods}
\alias{genFLStock,FLStock,FLQuant,FLQuant,missing-method}
\alias{genFLStock,FLStock,missing,FLQuant,FLQuant-method}
\alias{genFLStock,FLStock,FLQuant,missing,FLQuant-method}
\title{Methods to generate FLStock objects}
\usage{
genFLStock(object, R, C, F, ...)

\S4method{genFLStock}{FLStock,FLQuant,FLQuant,missing}(object, R, C, F,
  ...)

\S4method{genFLStock}{FLStock,missing,FLQuant,FLQuant}(object, R, C, F,
  ...)

\S4method{genFLStock}{FLStock,FLQuant,missing,FLQuant}(object, R, C, F,
  ...)
}
\arguments{
\item{object}{an FLStock}

\item{R}{an FLQuant with iterations or missing}

\item{C}{an FLQuant with iterations or missing}

\item{F}{an FLQuant with iterations or missing}

\item{...}{additional argument list that might not ever be used.}
}
\value{
an FLStock
}
\description{
This method computes the \code{FLStock} slots consistently with the information provided by the \code{FLQuant}. It requires two of the triplet R/C/F to compute the third consistent with Baranov and survival's equations.
}
