% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gen-methods.R
\name{getAcor}
\alias{getAcor}
\alias{getAcor,FLQuant-method}
\alias{getAcor-methods}
\title{compute log correlation matrix}
\usage{
getAcor(object, ...)
}
\arguments{
\item{object}{an \code{FLQuant} object}
}
\value{
a \code{FLQuant} object with the age correlation matrix
}
\description{
method to compute the log correlation matrix
}

