% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{methods}
\name{getTPL}
\alias{getTPL}
\title{Get TPL with ADMB code}
\usage{
getTPL(dir)
}
\arguments{
\item{dir}{folder where the a4a.tpl file will be copied to.}
}
\value{
file a4a.tpl
}
\description{
Function to get the a4a TPL file with ADMB code and copy into a specific folder.
}
\examples{
getTPL("myfolder")
}

