% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setupModel.R
\name{getX}
\alias{getX}
\alias{getX,formula-method}
\alias{getX-methods}
\title{Get model matrix}
\usage{
getX(object, ...)
}
\arguments{
\item{object}{a formula}

\item{...}{Additional argument list that might not ever be used.}
}
\value{
a matrix.
}
\description{
uses the user specified formula to build a model matrix
}
\note{
\code{getX} is intended to be used internally
}

