\name{grInvMod<-}
\alias{grInvMod<-}
\title{set growth inverse model}
\arguments{
  \item{object}{a \code{a4aGr} object}

  \item{value}{a \code{formula} object}
}
\value{
  a \code{a4aGr} object
}
\description{
  set method for \code{a4aGr} object's slot
  \code{grInvMod}.
}
\author{
  EJ \email{ernesto.jardim@jrc.ec.europa.eu}
}

