\name{grMod<-}
\alias{grMod<-}
\title{set growth model}
\arguments{
  \item{object}{a \code{a4aGr} object}

  \item{value}{a \code{formula} object}
}
\value{
  a \code{a4aGr} object
}
\description{
  set method for \code{a4aGr} object's slot \code{grMod}.
}
\author{
  EJ \email{ernesto.jardim@jrc.ec.europa.eu}
}

