% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hakeGSA7.idx}
\alias{hakeGSA7.idx}
\title{hakeGSA7.idx}
\format{
an \code{FLIndices} object
}
\source{
GFCM - STECF - MEDITS
}
\usage{
hakeGSA7.idx
}
\description{
Survey index for the Gulf of Lions Hake stock.
}
\author{
Chato Osio
}
