% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hakeGSA7}
\alias{hakeGSA7}
\title{hakeGSA7}
\format{an \code{FLStock} object}
\source{
GFCM - STECF
}
\usage{
hakeGSA7
}
\description{
Catch number, stocks weights, etc. for Gulf of Lions Hake (1998 - 2011).
}
\author{
Chato Osio
}

