\docType{methods}
\name{index.lvar}
\alias{index.lvar}
\alias{index.lvar,FLa4aFit-method}
\title{Hello}
\arguments{
  \item{object}{an FLQuant with iters}

  \item{...}{Additional argument list that might not ever
  be used.}
}
\value{
  an FLQuant
}
\description{
  Hello
}
\seealso{
  \code{\link{print}} and \code{\link{cat}}
}

