% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{index_cd_len}
\alias{index_cd_len}
\title{index_cd_len}
\format{an \code{FLIndex} object}
\usage{
index_cd_len
}
\description{
Survey abundance index for hake in the Gulf of Cadiz. Lenth frequencies.
}
\author{
Santiago Cervi\~no
}
