% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{index_pt_len}
\alias{index_pt_len}
\title{index_pt_len}
\format{
an \code{FLIndex} object
}
\usage{
index_pt_len
}
\description{
Survey abundance index for hake in Portuguese continental waters. Lenth frequencies.
}
\author{
Santiago Cervino
}
