% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{index_sp_len}
\alias{index_sp_len}
\title{index_sp_len}
\format{an \code{FLIndex} object}
\usage{
index_sp_len
}
\description{
Survey abundance index for hake in Northwest Spanish waters. Lenth frequencies.
}
\author{
Santiago Cervi\~no
}
