\name{is.empty}
\alias{is.empty}
\title{Assorted methods needed by FLa4a}
\usage{
  is.empty(object)
}
\arguments{
  \item{object}{an object}
}
\value{
  \code{TRUE} if object is of length 0
}
\description{
  Assorted methods needed by FLa4a
}
\examples{
is.empty(list())
is.empty(list(a=2))
}
\author{
  EJ \email{ernesto.jardim@jrc.ec.europa.eu}
}

