% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ma-methods.R
\name{ma}
\alias{ma}
\alias{ma,a4aFitSAs-method}
\alias{ma-methods}
\title{Model averaging (experimental)}
\usage{
ma(object, ...)
}
\arguments{
\item{object}{a \code{a4aFits} object with the fits to be averaged across}

\item{...}{Additional argument list that might not ever be used.}
}
\value{
a \code{FLStock} object with iterations defined by \code{nsim}
}
\description{
Method to average across a set of models. This is still experimental. Use with care.
}
\examples{
data(ple4)
data(ple4.indices)
f1 <- sca(ple4, ple4.indices, fmodel=~ factor(age) + s(year, k=20),
          qmodel=list(~ s(age, k = 4), ~ s(age, k = 4), ~ s(age, k = 3)),
          fit = "assessment")
f2 <- sca(ple4, ple4.indices, fmodel=~ factor(age) + s(year, k=20),
          qmodel=list(~ s(age, k = 4)+year, ~ s(age, k = 4), ~ s(age, k = 3)),
          fit = "assessment")
stock.sim <- ma(a4aFitSAs(list(f1=f1, f2=f2)), ple4, AIC, nsim = 100)
stks <- FLStocks(f1=ple4+f1, f2=ple4+f2, ma=stock.sim)
flqs <- lapply(stks, ssb)
flqs <- lapply(flqs, iterMedians)
xyplot(data~year, groups=qname, data=flqs, type="l")
plot(stks)
}

