\name{makePrediction}
\alias{makePrediction}
\title{Make predictions}
\usage{
  makePrediction(b.sim, X, e1, e2)
}
\arguments{
  \item{b.sim}{a \code{a4aFitSA} object}

  \item{X}{a design matrix}

  \item{e1}{whatever}

  \item{e2}{whatever}
}
\value{
  ??
}
\description{
  Method to make predictions
}

