% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a4aGr-methods.R
\docType{methods}
\name{mvrcop for a4aGr}
\alias{mvrcop for a4aGr}
\alias{mvrcop,numeric,a4aGr-method}
\title{mvrcop}
\usage{
\S4method{mvrcop}{numeric,a4aGr}(n = 1, mvdc, ...)
}
\arguments{
\item{n}{the number of iterations}

\item{mvdc}{the \code{a4aGr} object}

\item{...}{arguments to be passed to the rMvdc and copula methods}
}
\value{
an \code{FLModelSim} object with n groups of parameters
}
\description{
Method to generate multivariate parameters with user-defined copulas and marginals for \code{a4aGr} objects.
}
\examples{
mm <- matrix(NA, ncol=3, nrow=3)
diag(mm) <- c(50, 0.001,0.001)
mm[upper.tri(mm)] <- mm[lower.tri(mm)] <- c(0.1,0.01,0.00004)
vbObj <- a4aGr(grMod=~linf*(1-exp(-k*(t-t0))), grInvMod=~t0-1/k*log(1-len/linf), params=FLPar(linf=58.5, k=0.086, t0=0.001, units=c("cm","yr^-1","yr")), vcov=mm, distr="norm")
pars <- list(list(a=50, b=100, c=58.5), list(a=0.06, b=0.2, c=0.086), list(a=0, b=0.005, c=0.001))
#In the following, the third, fourth and fifth arguments refer to the copula,
#  while the final two arguments refer to the marginal distributions:
vbObj <- mvrcop(10000, vbObj, copula="archmCopula", family="clayton", param=2, margins="triangle", paramMargins=pars)
splom(data.frame(t(params(vbObj)@.Data)), pch=".")
}

