\docType{methods}
\name{niters for a4aGr}
\alias{niters for a4aGr}
\alias{niters,a4aGr-method}
\title{number of iterations}
\usage{
\S4method{niters}{a4aGr}(object)
}
\arguments{
  \item{object}{a \code{a4aGr} object}
}
\value{
a \code{numeric} object
}
\description{
method to extract from \code{a4aGr} objects the number of
iterations.
}

