\docType{methods}
\name{pars}
\alias{pars}
\alias{pars,a4aFitSA-method}
\title{Calculate the median accross iterations}
\arguments{
  \item{object}{an FLQuant with iters}

  \item{...}{Additional argument list that might not ever
  be used.}
}
\value{
  an FLQuant
}
\description{
  Calculate the median accross iterations

  Title
}
\examples{
data(ple4)
}
\seealso{
  \code{\link{print}} and \code{\link{cat}}
}

