% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLModelSimMethods.R
\name{pars2dim}
\alias{pars2dim}
\alias{pars2dim-methods}
\alias{pars2dim,FLModelSim-method}
\title{Check that the second dimension in params is "iter"}
\usage{
pars2dim(object)
}
\description{
Checks that the name of the second dimension in params is "iter". For internal use, not very interesting for users. It takes a \code{FLModelSim} object and returns a \code{logical}.
}
\examples{
pars2dim(FLModelSim())
}
