% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\docType{methods}
\name{plot,a4aFit,FLIndices-method}
\alias{plot,a4aFit,FLIndices-method}
\title{Plot of fitted indices-at-age}
\usage{
\S4method{plot}{a4aFit,FLIndices}(x, y, ...)
}
\arguments{
\item{x}{a \code{a4aFit} object with the fitted values}

\item{y}{a \code{FLIndices} object with the observed values}

\item{...}{Additional argument list that might not ever be used.}
}
\value{
a \code{plot} with fitted and observed indices-at-age
}
\description{
Method to plot fitted versus observed indices-at-age
}
\examples{
data(ple4)
data(ple4.index)
obj <- sca(ple4, FLIndices(ple4.index))
}

