% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/a4aFitresiduals-class.R
\docType{methods}
\name{plot,a4aFitResiduals,missing-method}
\alias{plot,a4aFitResiduals,missing-method}
\title{Plot of standardized log residuals}
\usage{
\S4method{plot}{a4aFitResiduals,missing}(x, y = missing, ...)
}
\arguments{
\item{x}{a \code{a4aFitResiduals} object with the standardized residuals}

\item{...}{Additional argument list that might not ever be used.}
}
\value{
a \code{plot} with stardardized log residuals
}
\description{
Method to plot scatterplot of standardized residuals
}
\examples{
data(ple4)
data(ple4.index)
obj <- sca(ple4, FLIndices(ple4.index))
}

