% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a4aFitCatchDiagn-class.R
\docType{methods}
\name{plot of catch residuals}
\alias{plot of catch residuals}
\alias{plot,a4aFitCatchDiagn,missing-method}
\title{Plot of aggregated catch standardized log residuals}
\usage{
\S4method{plot}{a4aFitCatchDiagn,missing}(x, y = missing, probs = c(0.1, 0.9), type = "all", ...)
}
\arguments{
\item{x}{an \code{a4aFit} object with the model fit}

\item{y}{the \code{FLStock} object used to fit the model}

\item{...}{additional argument list that might never be used}
}
\value{
a \code{plot} with stardardized log residuals
}
\description{
Method to produce scatterplots of aggregated catch residuals
}
\examples{
data(ple4)
data(ple4.index)
fit <- sca(ple4, ple4.index)
flqs <- computeCatchDiagnostics(fit, ple4)
plot(flqs)
}
