% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a4aFitMCMCs-class.R
\docType{methods}
\name{plot metrics of multiple fits}
\alias{plot metrics of multiple fits}
\alias{plot,a4aFitMCMCs,}
\alias{missing-method}
\title{Plot of metrics of multiple fits}
\usage{
\S4method{plot}{a4aFitMCMCs,missing}(x, y = missing, ...)
}
\arguments{
\item{x}{an \code{a4aFitMCMCs} object with multiple fits}

\item{y}{ignored}

\item{...}{additional argument list that might never be used}
}
\value{
a \code{plot} with fitting statistics
}
\description{
Method to plot fitting statistics of multiple fits, useful to compare fits.
}
\examples{
data(ple4)
data(ple4.index)
qmods <- list(list(~s(age, k=6)))
fmods = list()
for(i in 1:6) {
  fmods[[paste0(i)]] <-
    as.formula(
      paste0("~te(age, year, k = c(6,", i+14,"), bs = 'tp') + s(age, k = 6)")
    )
}
myFits <-
  FLa4a:::multisca(
    FLStocks(ple4), list(FLIndices(ple4.index)),
    fmodel = fmods, qmodel=qmods, fit="MCMC", mcmc=SCAMCMC()
  )
plot(myFits)
}
