% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{methods}
\name{plot for fitted indices-at-age}
\alias{plot for fitted indices-at-age}
\title{plot for fitted indices-at-age}
\usage{
\S4method{plot}{a4aFit,FLIndices}(x, y, ...)
}
\arguments{
\item{x}{an \code{a4aFit} object with the fitted values}

\item{y}{an \code{FLIndices} object with the observed values}

\item{...}{additional argument list that might never be used}
}
\value{
a \code{plot} with fitted and observed indices-at-age
}
\description{
Method to plot fitted versus observed indices-at-age.
}
\examples{
data(ple4)
data(ple4.index)
obj <- sca(ple4, FLIndices(ple4.index))
plot(obj, FLIndices(ple4.index))
}

