% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a4aGr-methods.R
\docType{methods}
\name{predict for a4aGr}
\alias{predict for a4aGr}
\alias{predict,a4aGr-method}
\title{predict for \code{a4aGr}}
\usage{
\S4method{predict}{a4aGr}(object, ...)
}
\arguments{
\item{object}{the \code{a4aGr} object}

\item{...}{arguments to be passed to the rMvdc and copula methods}
}
\value{
a \code{matrix} object with lengths or ages
}
\description{
Predicts ages or lengths using a growth class
}
\examples{
# Set up the a4aGr object and parameters for the marginals
mm <- matrix(NA, ncol=3, nrow=3)
diag(mm) <- c(50, 0.001,0.001)
mm[upper.tri(mm)] <- mm[lower.tri(mm)] <- c(0.1,0.01,0.00004)
md <- ~linf*(1-exp(-k*(t-t0)))
imd <- ~t0-1/k*log(1-len/linf)
prs <- FLPar(linf=58.5, k=0.086, t0=0.001, units=c("cm","yr^-1","yr"))
vbObj <- a4aGr(grMod=md, grInvMod=imd, params=prs, vcov=mm, distr="norm")
predict(vbObj, len=1:50+0.5)
predict(vbObj, t=1:20+0.5)
}
