% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a4aFitresiduals-class.R
\docType{methods}
\name{qqmath}
\alias{qqmath}
\alias{qqmath,a4aFitResiduals,missing-method}
\title{qqplot of standardized log residuals}
\usage{
qqmath(x, data, ...)

\S4method{qqmath}{a4aFitResiduals,missing}(x, data = missing, ...)
}
\arguments{
\item{x}{an \code{a4aFitResiduals} object with the standardized residuals}

\item{data}{ignored}

\item{...}{additional argument list that might never be used}
}
\value{
a \code{qqplot} with stardardized log residuals
}
\description{
Method to produce qqplots of standardized residuals
}
\examples{
data(ple4)
data(ple4.index)
obj <- sca(ple4, FLIndices(ple4.index))
flqs <- residuals(obj, ple4, FLIndices(idx=ple4.index))
qqmath(flqs)
}

