\name{quant2mat}
\alias{quant2mat}
\title{coerces FLPar into list for call}
\usage{
  quant2mat(x)
}
\arguments{
  \item{x}{An FLPar object}
}
\value{
  a list with each parameter as one element
}
\description{
  coerces FLPar into list for call
}
\author{
  EJ \email{ernesto.jardim@jrc.ec.europa.eu}
}

