% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{ra}
\alias{ra}
\alias{ra,a4aFit-method}
\alias{ra-methods}
\title{Retrospective analysis}
\usage{
ra(stock, indices, ...)
}
\arguments{
\item{stock}{A \code{FLStock} object.}

\item{indices}{A \code{FLIndices} object with the abundance and biomass indices.}

\item{...}{To pass the model specifications to fit the model.}

\item{n}{A \code{numeric} with the number of years to go back in the retrospectiva analysis.}
}
\value{
A \code{FLStocks} object with the \code{n} stocks resulting from removing one year and refitting the model.
}
\description{
Runs a retrospective analysis of the fit.
}
\note{
The default method will use the default submodels of \code{sca}.
}
\examples{
# load some data
data(ple4)
data(ple4.indices)
retro <- ra(ple4, ple4.indices, 5)
plot(retro)
}

