% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a4aM-methods.R
\docType{methods}
\name{range<-,a4aM,ANY,numeric-method}
\alias{range<-,a4aM,ANY,numeric-method}
\title{range for a4aM objects}
\usage{
\S4method{range}{a4aM,ANY,numeric}(x, i) <- value
}
\arguments{
\item{x}{an a4aM object}

\item{i}{the elements of range to be changed in a character vector}

\item{value}{a numeric vector with values}
}
\description{
Range method for a4aM objects
}
