% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/redfish-data.R
\docType{data}
\name{rfLen}
\alias{rfLen}
\title{redfish length data}
\format{an \code{FLStock}.}
\source{
Daniel Howell
}
\usage{
rfLen
}
\description{
Simulated length data for redfish.  Simulations were done using GADGET.
}
\author{
Ernesto Jardim
}

