% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rfTrawlJmp.idx}
\alias{rfTrawlJmp.idx}
\title{rfTrawlJmp.idx}
\format{an \code{FLIndex} object}
\usage{
rfTrawlJmp.idx
}
\description{
Trawl survey index for red fish, with a jump in catchability.
}
\author{
Daniel Howell
}

