\name{rngage<-}
\alias{rngage<-}
\title{age range replacement}
\arguments{
  \item{object}{a \code{FLComp} object}

  \item{value}{a \code{vector} with max and min age range}
}
\value{
  a \code{a4aM} object
}
\description{
  method to replace \code{FLComp} object's age range.
}
\author{
  EJ \email{ernesto.jardim@jrc.ec.europa.eu}
}

