% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCompMethods.R
\docType{methods}
\name{rngage}
\alias{rngage}
\alias{rngage,FLComp-method}
\alias{rngage-methods}
\alias{rngage<-}
\alias{rngage<-,FLComp,numeric-method}
\alias{rngage<--methods}
\title{age range extract and replacement}
\usage{
rngage(object, ...)

\S4method{rngage}{FLComp}(object)

rngage(object) <- value

\S4method{rngage}{FLComp,numeric}(object) <- value
}
\arguments{
\item{object}{object of relevant class (see signature of method)}

\item{...}{additional argument list that might never be used}

\item{value}{a \code{vector} with max and min age range to replace the object info}
}
\value{
a \code{vector} object when extracting, or an \code{FLComp} object when replacing
}
\description{
Methods to extract from \code{FLComp} objects the age range, or to replace its value.
}
\examples{
data(ple4)
data(ple4.index)
rngage(ple4)
rngage(ple4.index)
}

