\name{rnglen}
\alias{rnglen}
\alias{rnglen,FLComp-method}
\alias{rnglen-methods}
\alias{rnglen<-}
\alias{rnglen<-,FLComp,numeric-method}
\alias{rnglen<--methods}
\title{length range extract and replacement}
\arguments{
  \item{object}{a \code{FLComp} object}

  \item{value}{a \code{vector} with max and min length
  range to replace the object info.}
}
\value{
  a \code{vector} object when extracting or a \code{FLComp}
  object when replacing
}
\description{
  Methods to extract from \code{FLComp} objects the length
  range or replace its value.
}

