% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a4aM-class.R
\docType{methods}
\name{rngmbar}
\alias{rngmbar}
\alias{rngmbar,a4aM-method}
\alias{rngmbar-methods}
\alias{rngmbar<-}
\alias{rngmbar<-,a4aM-method}
\alias{rngmbar<--methods}
\title{M bar range extract and replacement}
\usage{
rngmbar(object, ...)

\S4method{rngmbar}{a4aM}(object)

rngmbar(object) <- value

\S4method{rngmbar}{a4aM}(object) <- value
}
\arguments{
\item{object}{object of relevant class (see signature of method)}

\item{...}{additional argument list that might never be used}

\item{value}{a \code{vector} with max and min age range to replace the object info}
}
\value{
a \code{vector} object when extracting or an \code{a4aM} object when replacing
}
\description{
Methods to extract from \code{a4aM} objects the M bar age range, or replace its value.
}
\examples{
mod1 <- FLModelSim(model=~exp(-age-0.5))
mod2 <- FLModelSim(model=~1.5*k, params=FLPar(k=0.4))
m1 <- a4aM(shape=mod1, level=mod2)
rngmbar(m1)<-c(1,5)
rngmbar(m1)
}

