% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCompMethods.R
\docType{methods}
\name{rngquant}
\alias{rngquant}
\alias{rngquant,FLComp-method}
\alias{rngquant-methods}
\alias{rngquant<-}
\alias{rngquant<-,FLComp,numeric-method}
\alias{rngquant<--methods}
\title{quant range extract and replacement}
\usage{
rngquant(object, ...)

\S4method{rngquant}{FLComp}(object)

rngquant(object) <- value

\S4method{rngquant}{FLComp,numeric}(object) <- value
}
\arguments{
\item{object}{object of relevant class (see signature of method)}

\item{...}{additional argument list that might never be used}

\item{value}{a \code{vector} with max and min quant range to replace the object info}
}
\value{
a \code{vector} object when extracting, or an \code{FLComp} object when replacing
}
\description{
Methods to extract from \code{FLComp} objects the quant range, or to replace its value.
}
\examples{
data(ple4.index)
data(rfLen)
rngquant(ple4.index)
rngquant(rfTrawl.idx)
}

