\name{simulate}
\alias{simulate}
\alias{simulate,a4aFitSA-method}
\alias{simulate,a4aStkParams-method}
\alias{simulate,SCAPars-method}
\alias{simulate,submodel-method}
\alias{simulate,submodels-method}
\title{Simulation methods for stock assessment fits}
\description{
  Simulation methods for stock assessment fits
}

