% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/southern_hake-data.R
\docType{data}
\name{southernHakeLen}
\alias{southernHakeLen}
\title{Southern hake length data}
\format{an \code{FLStockLen} and three \code{FLIndex} objects.}
\source{
Santiago Cervino
}
\usage{
southernHakeLen
}
\description{
Length based stock and three indices data for Southern hake.
}
\author{
Finlay Scott
}

