% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_submodels-class.R
\docType{class}
\name{stk_submodels}
\alias{stk_submodels}
\alias{a4aStkModel-class}
\alias{stk_submodels-methods}
\alias{m,stk_submodels-method}
\alias{wt,stk_submodels-method}
\alias{show,stk_submodels-method}
\alias{iter,stk_submodels-method}
\title{Stock submodels class}
\usage{
stk_submodels(object, ...)

\S4method{m}{stk_submodels}(object)

\S4method{wt}{stk_submodels}(object)

\S4method{show}{stk_submodels}(object)

\S4method{iter}{stk_submodels}(obj, it)
}
\arguments{
\item{obj}{the object to be subset}

\item{it}{iteration to be extracted}
}
\description{
Class definition (slots), constructors, accessors, replacement (when relevant) and common methods.
}
\section{Slot}{

\describe{
 \item{\code{fmodel}}{F submodel \code{formula}}
 \item{\code{n1model}}{first year N \code{formula}}
 \item{\code{srmodel}}{stock-recruitment submodel \code{formula}}
 \item{\code{m}}{natural mortality \code{FLQuant}}
 \item{\code{wt}}{stock weights \code{FLQuant}}
 \item{\code{units}}{data units \code{character}}
}
}

\section{Accessors}{

All slots in the class have accessor and replacement methods defined that
allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot.
A numeric vector can also be used when replacing FLQuant slots, and the
vector will be used to substitute the values in the slot, but not its other
attributes.
}

\section{Constructor}{

A construction method exists for this class that can take named arguments for
any of its slots. All slots are then created to match the requirements of the
class validity. If an unnamed \code{FLQuant} object is provided, this is used
for sizing, but not for populating any slot.
}

