% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodel_sim-class.R
\docType{class}
\name{submodel_sim}
\alias{submodel_sim}
\alias{submodel_sim-class}
\title{Submodel_sim class}
\description{
Class definition (slots), constructors, accessors, replacement (when relevant) and common methods.
}
\section{Slot}{

\describe{

 \item{\code{formula}}{\code{formula} describing the model}

 \item{\code{coefficients}}{\code{FLPar} with model parameters}

 \item{\code{vcov}}{\code{array} with variance covariance paramaters
                                 related to the variance model}

 \item{\code{centering}}{\code{numeric} value used for centering the data}

 \item{\code{distr}}{a character with the parameters' statistical
                     distribution; it must match a known distribution for
                     R (\emph{e.g.} "norm" for gaussian) so that \code{rnorm}
                     can be called}

 \item{\code{link}}{\code{function} value used for centering the data}

 \item{\code{linkinv}}{\code{function} value used for centering the data}

 \item{\code{covariates}}{\code{FLQuants} list of FLQuant objects holding}

}
}

