% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a4amse-is.R
\name{tac.is}
\alias{tac.is}
\title{Evaluate the chosen Management implementation function}
\usage{
tac.is(stk, ctrl, ay, nsqy = 3, delta_tac_max = NA, delta_tac_min = NA,
  tracking)
}
\arguments{
\item{stk}{The perceived stock.}

\item{ay}{The current year. The management control (e.g. TAC or effort) will be set in ay+1.}

\item{method}{Name of the chosen implementation function.}

\item{EFF}{Effort array (only used if effort management is being used).}

\item{EFF0}{Tracking array.}

\item{control}{The control object for the chosen management implementation function. A list of parameters.
TAC implementation function

Performs a short term forecast (STF) to hit the target F in year ay+1.
The resulting catch in year ay+1 is the TAC, i.e. the TAC that will result in Fbar = Ftarget.
The STF uses geometric mean recruitment. Fbar in the intermediate years (i.e. years between last data year and ay+1) are set as the mean of the last nsqy years.
The control argument is a list of parameters:
nsqy - number of years to average over to get Fbar for STF
delta_tac_min - constraint on the TAC
delta_tac_max - constraint on the TAC}

\item{stk}{The perceived FLStock.}

\item{imp_control}{A list with the elements: nsqy, delta_tac_min, delta_tac_max}

\item{ay}{The year for which the target F is set, based on the SSB in year (ay - control$ssb_lag).}

\item{EFF0}{The tracking array}

\item{EFF}{Not used by this function but may be used by the other implementation functions}
}
\description{
Evaluate the chosen HCR function using the current stock perception and a control.
For example, TAC control or effort control.
Returns a ctrl object for projecting the OM.
}
