\name{trend<-}
\alias{trend<-}
\title{trend replacement}
\arguments{
  \item{object}{a \code{a4aM} object}

  \item{value}{a \code{FLModelSim} object}
}
\value{
  a \code{a4aM} object
}
\description{
  replacement method for \code{a4aM} object's slot
  \code{trend}.
}
\author{
  EJ \email{ernesto.jardim@jrc.ec.europa.eu}
}

