% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCompMethods.R
\docType{methods}
\name{vecage}
\alias{vecage}
\alias{vecage,FLComp-method}
\alias{vecage-methods}
\title{age vector extract}
\usage{
vecage(object, ...)

\S4method{vecage}{FLComp}(object)
}
\arguments{
\item{object}{object of relevant class (see signature of method)}

\item{...}{additional argument list that might never be used}
}
\value{
a \code{vector} object
}
\description{
Method to extract from \code{FLComp} objects the vector of ages.
}
\examples{
data(ple4)
data(ple4.index)
vecage(ple4)
vecage(ple4.index)
}

