% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCompMethods.R
\docType{methods}
\name{veclen}
\alias{veclen}
\alias{veclen,FLComp-method}
\alias{veclen-methods}
\title{length vector}
\usage{
veclen(object, ...)

\S4method{veclen}{FLComp}(object)
}
\arguments{
\item{object}{object of relevant class (see signature of method)}

\item{...}{additional argument list that might never be used}
}
\value{
a \code{vector} object
}
\description{
Method to extract from \code{FLComp} objects the vector of lengths.
}
\examples{
data(rfLen)
veclen(rfTrawl.idx)
}

