\name{vecmbar}
\alias{vecmbar}
\alias{vecmbar,a4aM-method}
\alias{vecmbar-methods}
\title{M age vector}
\usage{
vecmbar(object, ...)
}
\arguments{
  \item{object}{a \code{a4aM} object}
}
\value{
a \code{vector} object
}
\description{
method to extract from \code{a4aM} objects the vector of
ages for the mean M.
}

