% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCompMethods.R
\name{vecquant}
\alias{vecquant}
\alias{vecquant-methods}
\alias{vecquant,FLComp-method}
\title{length vector}
\usage{
vecquant(object, ...)
}
\arguments{
\item{object}{a \code{FLComp} object}
}
\value{
a \code{vector} object
}
\description{
method to extract from \code{FLComp} objects the vector of lengths.
}
\examples{
data(ple4.index)
data(rfLen)
vecquant(ple4.index)
vecquant(rfTrawl.idx)
}
