% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLCompMethods.R
\name{vecyear}
\alias{vecyear}
\alias{vecyear,FLComp-method}
\alias{vecyear-methods}
\title{year vector extract}
\usage{
vecyear(object, ...)
}
\arguments{
\item{object}{a \code{FLComp} object}
}
\value{
a \code{vector} object
}
\description{
method to extract from \code{FLComp} objects the vector of years.
}

