% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fittingFunctions.R
\docType{methods}
\name{a4a}
\alias{a4a}
\title{a4a}
\usage{
a4a(...)
}
\arguments{
\item{...}{additional argument list that might never be used}
}
\description{
This was the old stock assessment framework user's interface. It was replaced by \code{sca} or the advanced version \code{a4aSCA}.
}
