% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a4aFit-class.R
\docType{class}
\name{a4aFit-class}
\alias{a4aFit-class}
\alias{stock.n,a4aFit-method}
\alias{harvest,a4aFit,ANY-method}
\alias{catch.n,a4aFit-method}
\alias{index,a4aFit-method}
\alias{coef,a4aFit-method}
\alias{vcov,a4aFit-method}
\alias{fitSumm}
\alias{fitSumm,a4aFit-method}
\alias{logLik,a4aFit-method}
\alias{show,a4aFit-method}
\alias{print,a4aFit-method}
\title{The a4aFit class}
\usage{
\S4method{stock.n}{a4aFit}(object)

\S4method{harvest}{a4aFit,ANY}(object, catch, ...)

\S4method{catch.n}{a4aFit}(object)

\S4method{index}{a4aFit}(object, ...)

\S4method{coef}{a4aFit}(object, ...)

\S4method{vcov}{a4aFit}(object, ...)

fitSumm(object, ...)

\S4method{fitSumm}{a4aFit}(object)

\S4method{logLik}{a4aFit}(object, ...)

\S4method{show}{a4aFit}(object)

\S4method{print}{a4aFit}(x, ...)
}
\arguments{
\item{object}{an \code{a4aFit}.}

\item{catch, ...}{not used.}

\item{x}{an \code{a4aFit}.}
}
\description{
Results of a statistical catch-at-age fit with \code{\link[=sca]{sca()}}.
}
\details{
\code{logLik()} returns the maximised log-likelihood (one value per
iteration), so \code{AIC()} and \code{BIC()} work on fits.
}
\section{Slots}{

\describe{
\item{\code{call}}{the call to \code{\link[=sca]{sca()}}.}

\item{\code{stock.n,harvest,catch.n}}{estimated numbers, fishing mortality and catch at age.}

\item{\code{index}}{fitted indices.}

\item{\code{fitSumm}}{fit summary (statistic x iter): number of parameters,
negative log-likelihood, maximum gradient, number of observations,
convergence flag and likelihood components.}

\item{\code{coefficients}}{parameter estimates (\code{FLPar}, params x iter).}

\item{\code{vcov}}{parameter covariance (params x params x iter); \code{NA} when
\code{fit = "MP"}.}

\item{\code{centering}}{log-scale centering applied to each fleet (\code{FLPar}).}

\item{\code{models}}{the submodel formulas.}
}}

