% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a4aFitMCMC-class.R
\docType{class}
\name{a4aFitMCMC-class}
\alias{a4aFitMCMC}
\alias{a4aFitMCMC,a4aFitSA-method}
\alias{a4aFitMCMC,missing-method}
\alias{a4aFitMCMC-class}
\alias{a4aFitMCMC-methods}
\alias{a4aFitSA,a4aFitMCMC-method}
\alias{as.mcmc}
\alias{as.mcmc,a4aFitMCMC-method}
\alias{as.mcmc-methods}
\title{S4 class \code{a4aFitMCMC}}
\usage{
a4aFitMCMC(object, ...)

\S4method{a4aFitMCMC}{missing}(object, ...)

\S4method{a4aFitSA}{a4aFitMCMC}(object, ...)

\S4method{a4aFitMCMC}{a4aFitSA}(object, ...)

as.mcmc(x, ...)

\S4method{as.mcmc}{a4aFitMCMC}(x, ...)
}
\arguments{
\item{object}{object of relevant class (see signature of method)}

\item{...}{additional argument list that might never be used}

\item{x}{an object to be coerced into mcmc}
}
\description{
The \code{a4aFitMCMC} class extends \code{a4aFitSA} to store information about the MCMC run.
}
\section{Slots}{

\describe{
   \item{SCAMCMC}{An object of class \code{SCAMCMC} with information about the MCMC run}
}
}

\section{Accessors}{

All slots in the class have accessor and replacement methods defined that
allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot.
A numeric vector can also be used when replacing FLQuant slots, and the
vector will be used to substitute the values in the slot, but not its other
attributes.
blah blah
}

\section{Constructor}{

A construction method exists for this class that can take named arguments for
any of its slots. All slots are then created to match the requirements of the
class validity. If an unnamed \code{FLQuant} object is provided, this is used
for sizing, but not for populating any slot.
}
\examples{
data(ple4)
data(ple4.index)

obj <- sca(stock=ple4, indices=FLIndices(ple4.index), fit="assessment")
obj

slotNames(obj)
clock(obj)
fitSumm(obj)

flq <- stock.n(obj)
is(flq)
flq <- index(obj)
is(flq)

logLik(obj)
AIC(obj)
BIC(obj)

is(pars(obj))
}

