% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/a4aFitresiduals-class.R
\docType{class}
\name{a4aFitResiduals-class}
\alias{a4aFitResiduals}
\alias{a4aFitResiduals-class}
\alias{a4aFitResiduals-methods}
\alias{residuals,a4aFit-method}
\title{S4 class \code{a4aFitResiduals}}
\usage{
\S4method{residuals}{a4aFit}(object, stock, indices, ...)
}
\description{
The \code{a4aFitResiduals} class extends \code{FLQuants} to store residuals of the a4a stock assessment fit. By default these should be log residuals of catches and indices.
}
\examples{
data(ple4)
data(ple4.index)
obj <- sca(ple4, FLIndices(ple4.index))
}

