% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a4aFitresiduals-class.R
\docType{class}
\name{a4aFitResiduals-class}
\alias{a4aFitResiduals-class}
\alias{residuals,a4aFit-method}
\alias{a4aFitResiduals}
\alias{a4aFitResiduals-methods}
\title{S4 class \code{a4aFitResiduals}}
\usage{
\S4method{residuals}{a4aFit}(object, stock, indices, ...)
}
\arguments{
\item{object}{object of relevant class (see signature of method)}

\item{stock}{\code{FLStock} object used to fit the model}

\item{indices}{\code{FLIndices} object used to fit the model}

\item{...}{additional argument list that might never be used}
}
\description{
The \code{a4aFitResiduals} class extends \code{FLQuants} to store residuals of the a4a stock assessment fit. By default, these should be log residuals of catches and indices.
}
\examples{
data(ple4)
data(ple4.index)
obj <- sca(ple4, ple4.index)
flqs <- residuals(obj, ple4, FLIndices(idx=ple4.index))
}
