% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fittingFunctions.R
\docType{methods}
\name{a4aInternal}
\alias{a4aInternal}
\title{Stock assessment model advanced method}
\usage{
a4aInternal(stock, indices, fmodel = ~s(age, k = 3) + factor(year),
  qmodel = lapply(seq(length(indices)), function(i) ~1),
  srmodel = ~factor(year), n1model = ~factor(age),
  vmodel = lapply(seq(length(indices) + 1), function(i) ~1),
  covar = missing, wkdir = missing, verbose = FALSE, fit = "assessment",
  center = TRUE, mcmc = missing, useADMB = TRUE)
}
\arguments{
\item{stock}{an FLStock object containing catch and stock information}

\item{indices}{an FLIndices object containing survey indices}

\item{fmodel}{a formula object depicting the model for log fishing mortality at age}

\item{qmodel}{a list of formula objects depicting the models for log survey catchability at age}

\item{srmodel}{a formula object depicting the model for log recruitment}

\item{n1model}{a formula object depicting the model for the first year of catch data}

\item{vmodel}{a list of formula objects depicting the models for log survey and log fishing mortality variance}

\item{covar}{a list with covariates}

\item{wkdir}{used to set a working directory for the admb optimiser.  If wkdir is set all admb files are saved to this folder otherwise they are deleted.}

\item{verbose}{if true admb fitting information is printed to the screen}

\item{fit}{character with type of fit: 'MP' or 'assessment'; the former doesn't require the hessian to be computed, while the latter does.}

\item{center}{\code{logical} specifying whether data is centered before estimating or not}

\item{mcmc}{\code{SCAMCMC} specifying parameters for the ADMB MCMC run, check ADMB manual for detailed description}

\item{useADMB}{if FALSE use TMB if TRUE (deafault) use ADMB (note MCMC cannot be used with TMB)}
}
\value{
an \code{a4aFit} object if fit is "MP" or an \code{a4aFitSA} if fit is "assessment"
}
\description{
The advanced user interface to the a4a fitting routine.
}
