% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fittingFunctions.R
\docType{methods}
\name{a4aInternal}
\alias{a4aInternal}
\title{Stock assessment model advanced method}
\usage{
a4aInternal(stock, indices, fmodel = defaultFmod(stock),
  qmodel = defaultQmod(indices), srmodel = defaultSRmod(stock),
  n1model = defaultN1mod(stock), vmodel = defaultVmod(stock, indices),
  covar = missing, wkdir = missing, verbose = FALSE,
  fit = "assessment", center = TRUE, mcmc = missing)
}
\arguments{
\item{stock}{an FLStock object containing catch and stock information}

\item{indices}{an FLIndices object containing survey indices}

\item{fmodel}{a formula object depicting the model for log fishing mortality at age}

\item{qmodel}{a list of formula objects depicting the models for log survey catchability at age}

\item{srmodel}{a formula object depicting the model for log recruitment}

\item{n1model}{a formula object depicting the model for the first year of catch data}

\item{vmodel}{a list of formula objects depicting the models for log survey and log fishing mortality variance}

\item{covar}{a list with covariates}

\item{wkdir}{used to set a working directory for the admb optimiser.  If wkdir is set all admb files are saved to this folder otherwise they are deleted.}

\item{verbose}{if true admb fitting information is printed to the screen}

\item{fit}{character with type of fit: 'MP' or 'assessment'; the former doesn't require the hessian to be computed, while the latter does.}

\item{center}{\code{logical} specifying whether data is centered before estimating or not}

\item{mcmc}{\code{SCAMCMC} specifying parameters for the ADMB MCMC run, check ADMB manual for detailed description}
}
\value{
an \code{a4aFit} object if fit is "MP" or an \code{a4aFitSA} if fit is "assessment"
}
\description{
The advanced user interface to the a4a fitting routine.
}
