% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a4aStkParams-class.R
\docType{class}
\name{a4aStkParams}
\alias{a4aStkParams}
\alias{a4aStkParams-class}
\alias{a4aStkParams}
\alias{a4aStkParams-methods}
\alias{a4aStkParams,missing-method}
\alias{m,a4aStkParams-method}
\alias{wt,a4aStkParams-method}
\alias{mat,a4aStkParams-method}
\alias{fMod}
\alias{fMod-methods}
\alias{fMod,a4aStkParams-method}
\alias{fMod<-}
\alias{fMod<--methods}
\alias{fMod<-,a4aStkParams,formula-method}
\alias{n1Mod}
\alias{n1Mod-methods}
\alias{n1Mod,a4aStkParams-method}
\alias{n1Mod<-}
\alias{n1Mod<--methods}
\alias{n1Mod<-,a4aStkParams,formula-method}
\alias{srMod}
\alias{rMod-methods}
\alias{srMod,a4aStkParams-method}
\alias{srMod<-}
\alias{srMod<--methods}
\alias{srMod<-,a4aStkParams,formula-method}
\alias{params,a4aStkParams-method}
\alias{params<-,a4aStkParams,FLPar-method}
\alias{coefficients}
\alias{coefficients-methods}
\alias{coefficients,a4aStkParams-method}
\alias{coefficients<-,a4aStkParams,FLPar-method}
\alias{distr,a4aStkParams-method}
\alias{distr<-,a4aStkParams,character-method}
\alias{vcov,a4aStkParams-method}
\alias{vcov<-,a4aStkParams,array-method}
\alias{propagate,a4aStkParams-method}
\alias{iter,a4aStkParams-method}
\title{Stock parameters class}
\usage{
a4aStkParams(object, ...)

\S4method{a4aStkParams}{missing}(object, ...)

\S4method{m}{a4aStkParams}(object)

\S4method{wt}{a4aStkParams}(object)

\S4method{mat}{a4aStkParams}(object)

fMod(object, ...)

\S4method{fMod}{a4aStkParams}(object)

fMod(object) <- value

\S4method{fMod}{a4aStkParams,formula}(object) <- value

n1Mod(object, ...)

\S4method{n1Mod}{a4aStkParams}(object)

n1Mod(object) <- value

\S4method{n1Mod}{a4aStkParams,formula}(object) <- value

srMod(object, ...)

\S4method{srMod}{a4aStkParams}(object)

srMod(object) <- value

\S4method{srMod}{a4aStkParams,formula}(object) <- value

\S4method{params}{a4aStkParams}(object)

\S4method{params}{a4aStkParams,FLPar}(object) <- value

coefficients(object, ...)

\S4method{coefficients}{a4aStkParams}(object)

\S4method{coefficients}{a4aStkParams,FLPar}(object) <- value

\S4method{distr}{a4aStkParams}(object)

\S4method{distr}{a4aStkParams,character}(object) <- value

\S4method{vcov}{a4aStkParams}(object)

\S4method{vcov}{a4aStkParams,array}(object) <- value

\S4method{propagate}{a4aStkParams}(object, iter, fill.iter = TRUE)

\S4method{iter}{a4aStkParams}(obj, it)
}
\arguments{
\item{object}{object of relevant class (see signature of method)}

\item{...}{additional argument list that might never be used}

\item{value}{the new object}

\item{iter}{the number of iterations to create}

\item{fill.iter}{should the new iterations be filled with values (TRUE) or NAs (FALSE)}

\item{obj}{the object to be subset}

\item{it}{iteration to be extracted}
}
\description{
Class definition (slots), constructors, accessors, replacement (when relevant) and common methods.
}
\section{Slot}{

\describe{
\item{\code{fMod}}{F submodel \code{formula}}
\item{\code{n1Mod}}{first year N \code{formula}}
\item{\code{srMod}}{stock-recruitment submodel \code{formula}}
\item{\code{params}}{\code{FLPar} with parameters}
\item{\code{vcov}}{\code{array} with variance-covariance}
\item{\code{centering}}{centering values \code{numeric}}
\item{\code{distr}}{statistical distribution \code{character}}
\item{\code{m}}{natural mortality \code{FLQuant}}
\item{\code{units}}{data units \code{character}}
}
}

\section{Accessors}{

All slots in the class have accessor and replacement methods defined that
allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot.
A numeric vector can also be used when replacing FLQuant slots, and the
vector will be used to substitute the values in the slot, but not its other
attributes.
}

\section{Constructor}{

A construction method exists for this class that can take named arguments for
any of its slots. All slots are then created to match the requirements of the
class validity. If an unnamed \code{FLQuant} object is provided, this is used
for sizing, but not for populating any slot.
}

