\docType{methods}
\name{+}
\alias{+}
\alias{+,FLIndices,SCAPars-method}
\alias{+,FLIndices,a4aFit-method}
\alias{+,FLIndices,a4aFitSA-method}
\alias{+,FLStock,SCAPars-method}
\alias{+,FLStock,a4aFit-method}
\alias{+,FLStock,a4aFitSA-method}
\title{+ methods}
\usage{
\S4method{+}{FLStock,a4aFit}(e1, e2)

\S4method{+}{FLStock,a4aFitSA}(e1, e2)

\S4method{+}{FLStock,SCAPars}(e1, e2)

\S4method{+}{FLIndices,a4aFit}(e1, e2)

\S4method{+}{FLIndices,a4aFitSA}(e1, e2)

\S4method{+}{FLIndices,SCAPars}(e1, e2)
}
\description{
+ methods
}

