% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getX-methods.R, R/fittingFunctions.R
\name{breakpts}
\alias{breakpts}
\alias{breakpts-methods}
\alias{breakpts,numeric-method}
\title{Breakpoints}
\usage{
breakpts(var, ...)

\S4method{breakpts}{numeric}(var, breaks, ...)

breakpts(var, ...)

\S4method{breakpts}{numeric}(var, breaks, ...)
}
\arguments{
\item{var}{a \code{numeric} object that defines the variable to be "broken"}

\item{...}{additional argument list that might never be used}

\item{breaks}{a \code{numeric} object that defines the breakpoints}
}
\value{
a \code{factor} with levels according to the defined breaks

a \code{factor} with levels according to the defined breaks
}
\description{
Method to set breakpoints in submodels

Method to set breakpoints in submodels
}
