% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fittingFunctions.R
\name{breakpts}
\alias{breakpts}
\alias{breakpts,numeric-method}
\alias{breakpts-methods}
\title{Breakpoints}
\usage{
breakpts(var, ...)
}
\arguments{
\item{var}{a \code{numeric} object that defines the variable to be "broken"}

\item{breaks}{a \code{numeric} object that defines the breakpoints}
}
\value{
a \code{factor} with levels according to the defined breaks
}
\description{
Method to set breakpoints in submodels
}

