% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a4aFitresiduals-class.R
\docType{methods}
\name{bubbles}
\alias{bubbles}
\alias{bubbles,a4aFitResiduals,missing-method}
\title{Bubbles plot of standardized log residuals}
\usage{
\S4method{bubbles}{a4aFitResiduals,missing}(x, data = missing, ...)
}
\arguments{
\item{x}{an \code{a4aFitResiduals} object with the standardized residuals}

\item{data}{ignored}

\item{...}{additional argument list that might never be used}
}
\value{
a \code{bubbles} plot with stardardized log residuals
}
\description{
Method to produce bubble plots of standardized residuals
}
\examples{
data(ple4)
data(ple4.index)
obj <- sca(ple4, FLIndices(ple4.index))
flqs <- residuals(obj, ple4, FLIndices(idx=ple4.index))
bubbles(flqs)
}

