% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodel-class.R, R/coef-methods.R
\docType{methods}
\name{formula<-}
\alias{formula<-}
\alias{coef<-,a4aFitSA-methods}
\alias{formula<-,submodel,formula-method}
\alias{coef}
\alias{coef,a4aFitSA-methods}
\alias{coef,a4aFitSA-method}
\alias{coef,SCAPars-method}
\alias{coef,a4aStkParams-method}
\alias{coef,submodels-method}
\alias{coef,submodel-method}
\alias{coef<-}
\alias{coef<-,a4aFitSA-methods}
\alias{coef<-,a4aFitSA,numeric-method}
\alias{coef<-,SCAPars,numeric-method}
\alias{coef<-,a4aStkParams,numeric-method}
\alias{coef<-,submodels,numeric-method}
\alias{coef<-,submodel,numeric-method}
\alias{coef<-,submodel,FLPar-method}
\alias{coef<-,a4aStkParams,FLPar-method}
\alias{coef<-,a4aStkParams,matrix-method}
\title{coefficients extract and replacement}
\usage{
formula(object) <- value

\S4method{formula}{submodel,formula}(object) <- value

coef(object, ...)

\S4method{coef}{a4aFitSA}(object)

\S4method{coef}{SCAPars}(object)

\S4method{coef}{a4aStkParams}(object)

\S4method{coef}{submodels}(object)

\S4method{coef}{submodel}(object)

coef(object, ...) <- value

\S4method{coef}{a4aFitSA,numeric}(object, ...) <- value

\S4method{coef}{SCAPars,numeric}(object, ...) <- value

\S4method{coef}{a4aStkParams,numeric}(object, ...) <- value

\S4method{coef}{submodels,numeric}(object, ...) <- value

\S4method{coef}{submodel,numeric}(object, ...) <- value

\S4method{coef}{submodel,FLPar}(object, ...) <- value

\S4method{coef}{a4aStkParams,FLPar}(object, ...) <- value

\S4method{coef}{a4aStkParams,matrix}(object, ...) <- value
}
\arguments{
\item{object}{object of relevant class (see signature of method)}

\item{value}{the new object}

\item{...}{additional argument list that might never be used}

\item{value}{the new object}
}
\description{
Methods to extract and replace the model coefficients.
}
