% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef-methods.R
\docType{methods}
\name{coef}
\alias{coef}
\alias{coef,SCAPars-method}
\alias{coef,a4aFitSA-method}
\alias{coef,a4aFitSA-methods}
\alias{coef,a4aStkParams-method}
\alias{coef,submodel-method}
\alias{coef,submodels-method}
\alias{coef<-}
\alias{coef<-,SCAPars,numeric-method}
\alias{coef<-,a4aFitSA,numeric-method}
\alias{coef<-,a4aFitSA-methods}
\alias{coef<-,a4aStkParams,numeric-method}
\alias{coef<-,submodel,numeric-method}
\alias{coef<-,submodels,numeric-method}
\title{coefficients extract and replacement}
\usage{
coef(object, ...)

\S4method{coef}{a4aFitSA}(object)

\S4method{coef}{SCAPars}(object)

\S4method{coef}{a4aStkParams}(object)

\S4method{coef}{submodels}(object)

\S4method{coef}{submodel}(object)

coef(object, ...) <- value

\S4method{coef}{a4aFitSA,numeric}(object, ...) <- value

\S4method{coef}{SCAPars,numeric}(object, ...) <- value

\S4method{coef}{a4aStkParams,numeric}(object, ...) <- value

\S4method{coef}{submodels,numeric}(object, ...) <- value

\S4method{coef}{submodel,numeric}(object, ...) <- value
}
\arguments{
\item{object}{object of relevant class (see signature of method)}

\item{...}{additional argument list that might never be used}

\item{value}{the new object}
}
\description{
Methods to extract and replace the model coefficients.
}

