% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fittingFunctions.R
\docType{methods}
\name{defaultSubModels}
\alias{defaultSubModels}
\alias{defaultFmod}
\alias{defaultQmod}
\alias{defaultN1mod}
\alias{defaultVmod}
\alias{defaultSRmod}
\title{Default sub-models}
\usage{
defaultFmod(stock, dfm = c(0.5, 0.7))

defaultQmod(indices, dfm = 0.6)

defaultN1mod(stock)

defaultVmod(stock, indices)

defaultSRmod(stock)
}
\arguments{
\item{stock}{an FLStock object}

\item{dfm}{numeric vector with the data points fraction to be used to set the spline ks.}

\item{indices}{an FLIndices object}
}
\value{
a FLStock object
}
\description{
Methods to create formulas for sub-models. The sub-models are set automagically using defaults.
}
