% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gen-methods.R
\docType{methods}
\name{genFLIndex}
\alias{genFLIndex}
\alias{genFLIndex,FLQuant-method}
\alias{genFLIndex-methods}
\title{Methods to generate FLIndex objects}
\usage{
genFLIndex(object, ...)

\S4method{genFLIndex}{FLQuant}(object, cv = 0.2, niter = 250)
}
\arguments{
\item{object}{an \code{FLIndex} object}

\item{...}{additional argument list that might not ever be used.}

\item{cv}{the coefficient of variation}

\item{niter}{the number of iterations to be generated}
}
\value{
an FLIndex
}
\description{
This method produces an \code{FLIndex} object by using the \code{genFLQuant} method.
}

