\docType{methods}
\name{genFLIndex}
\alias{genFLIndex}
\alias{genFLIndex,FLQuant-method}
\alias{genFLIndex-methods}
\title{Methods to generate FLIndex objects}
\arguments{
  \item{object}{an FLIndex}

  \item{cv}{the coefficient of variation}

  \item{niter}{number of iterations to be generated}

  \item{...}{Additional argument list that might not ever
  be used.}
}
\value{
  an FLIndex
}
\description{
  This method generates a \code{FLIndex} by generating the
  index using \code{genFLQuant} methods.
}

