% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-methods.R
\docType{methods}
\name{genFLQuant}
\alias{genFLQuant}
\alias{genFLQuant-methods}
\alias{genFLQuant,FLQuant-method}
\title{Methods to generate FLQuant objects}
\usage{
genFLQuant(object, ...)

\S4method{genFLQuant}{FLQuant}(object, cv = 0.2, method = "ac",
  niter = 250)
}
\arguments{
\item{object}{an FLQuant}

\item{...}{additional argument list that might not ever be used.}

\item{cv}{the coefficient of variation}

\item{method}{the method used to compute the correlation matrix; for now only "ac" (autocorrelation) is implemented}

\item{niter}{the number of iterations to be generated}
}
\value{
an FLQuant
}
\description{
This method uses the quant log-correlation matrix of the \code{FLQuant} object and generates a new \code{FLQuant} using a lognormal multivariate distribution.
}
\examples{
data(ple4)
sim.F <- genFLQuant(harvest(ple4))
}
