% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gen-methods.R
\docType{methods}
\name{genFLQuant}
\alias{genFLQuant}
\alias{genFLQuant,FLQuant-method}
\alias{genFLQuant-methods}
\title{Methods to generate FLQuant objects}
\usage{
genFLQuant(object, ...)

\S4method{genFLQuant}{FLQuant}(object, cv = 0.2, method = "ac",
  niter = 250)
}
\arguments{
\item{object}{an FLQuant}

\item{cv}{the coefficient of variation}

\item{method}{the ethod used to compute the correlation matrix, for now only "ac" - autocorrelation}

\item{niter}{number of iterations to be generated}
}
\value{
an FLQuant
}
\description{
This method uses the age correlation matrix of the \code{FLQuant} and generates a new \code{FLQuant} using a lognormal multivariate distribution
}
\examples{
data(ple4)
sim.F <- genFLQuant(harvest(ple4))
}

