% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-methods.R
\docType{methods}
\name{getAcor}
\alias{getAcor}
\alias{getAcor,FLQuant-method}
\alias{getAcor-methods}
\title{compute log-correlation matrix}
\usage{
getAcor(object, ...)

\S4method{getAcor}{FLQuant}(object, ...)
}
\arguments{
\item{object}{object of relevant class (see signature of method)}

\item{...}{additional argument list that might never be used}
}
\value{
an \code{FLQuant} object with a quant log-correlation matrix
}
\description{
Method to compute the log-correlation matrix for the first dimension (\code{quant}) of the \code{FLQuant} object.
}
\examples{
data(ple4)
getAcor(harvest(ple4))
}

