\name{getAcor}
\alias{getAcor}
\alias{getAcor,FLQuant-method}
\alias{getAcor-methods}
\title{compute log correlation matrix}
\arguments{
  \item{object}{an \code{FLQuant} object}
}
\value{
  a \code{FLQuant} object with the age correlation matrix
}
\description{
  method to compute the log correlation matrix
}

