% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getADMBHessian}
\alias{getADMBCovariance}
\alias{getADMBHessian}
\title{Get ADMB Hessian}
\usage{
getADMBHessian(wkdir)

getADMBCovariance(wkdir)
}
\arguments{
\item{wkdir}{the location of the admb output}
}
\value{
a list with the following elements
}
\description{
Reads the hessian file from any ADMB fit.  Used here with the a4a model.
}
\note{
\code{getADMBHessian} is intended to be used internally
}
\examples{
# load some data
data(ple4)
data(ple4.indices)
# choose a working directory
wkdir <- tempfile()
# do an 'assessment' fit with default settings (not recomended!) and keep results in wkdir
fit <- a4aSCA(stock=ple4,indices=ple4.indices,wkdir=wkdir,fit="assessment")
hessInfo <- getADMBHessian(wkdir)
str(hessInfo)
# calculate covariance matrix
Sigma <- solve(hessInfo$hes)
}

