% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a4aGr-methods.R
\docType{methods}
\name{getK}
\alias{getK}
\alias{getK,a4aGr-method}
\alias{getK-methods}
\title{Get K}
\usage{
getK(object, ...)

\S4method{getK}{a4aGr}(object)
}
\arguments{
\item{object}{an \code{a4aGr} object}

\item{...}{additional argument list that might never be used}
}
\value{
a \code{vector} with K values
}
\description{
Method to get values of the growth parameter K
}
\examples{
mm <- matrix(NA, ncol=3, nrow=3)
diag(mm) <- c(50, 0.001,0.001)
mm[upper.tri(mm)] <- mm[lower.tri(mm)] <- c(0.1,0.01,0.00004)
md <- ~linf*(1-exp(-k*(t-t0)))
imd <- ~t0-1/k*log(1-len/linf)
prs <- FLPar(linf=58.5, k=0.086, t0=0.001, units=c("cm","yr^-1","yr"))
vbObj <- a4aGr(grMod=md, grInvMod=imd, params=prs, vcov=mm, distr="norm")
vbObj <- mvrnorm(100,vbObj)
getK(vbObj)
}

