% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getX-methods.R
\name{getX}
\alias{getX}
\alias{getX-methods}
\alias{getX,submodels-method}
\alias{getX,submodel-method}
\alias{getX,submodel_sim-method}
\alias{getX,formula-method}
\title{Get model matrix}
\usage{
getX(object, ...)

\S4method{getX}{submodels}(object)

\S4method{getX}{submodel}(object)

\S4method{getX}{submodel_sim}(object)

\S4method{getX}{formula}(object, df, check = TRUE)
}
\arguments{
\item{object}{object of relevant class (see signature of method)}

\item{...}{additional argument list that might never be used}

\item{df}{the data.frame to build the model matrix against.}
}
\value{
a matrix.
}
\description{
Uses the user-specified formula to build a model matrix.
}
\note{
\code{getX} is intended to be used internally
}
