% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupModel.R
\docType{methods}
\name{getX}
\alias{getX}
\alias{getX,formula-method}
\alias{getX-methods}
\title{Get model matrix}
\usage{
getX(object, ...)

\S4method{getX}{formula}(object, df, newdf = df)
}
\arguments{
\item{object}{object of relevant class (see signature of method)}

\item{...}{additional argument list that might never be used}

\item{df}{the data.frame to build the model matrix against.}

\item{newdf}{the data.frame to create the model matrix for.}
}
\value{
a matrix.
}
\description{
Uses the user-specified formula to build a model matrix.
}
\note{
\code{getX} is intended to be used internally
}

