\name{getX}
\alias{getX}
\alias{getX,a4aFitSA-method}
\title{build model matrix}
\arguments{
  \item{formula}{a formula object}

  \item{df}{the data.frame to build the model matrix
  against}
}
\value{
  a matrix.
}
\description{
  uses the user specified formula to build a model matrix
}
\note{
  \code{getX} is intended to be used internally
}
\author{
  Colin Millar \email{colin.millar@jrc.ec.europa.eu}
}

