% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupModel.R
\name{getX}
\alias{getX}
\alias{getX-methods}
\alias{getX,formula-method}
\title{Get model matrix}
\usage{
getX(object, ...)
}
\arguments{
\item{formula}{a formula object}

\item{df}{the data.frame to build the model matrix against}

\item{df}{the data.frame to create the model matrix for.}
}
\value{
a matrix.
}
\description{
Uses the user-specified formula to build a model matrix.
}
\note{
\code{getX} is intended to be used internally
}
