% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma-methods.R
\docType{methods}
\name{ma}
\alias{ma}
\alias{ma,a4aFitSAs-method}
\alias{ma-methods}
\title{Model averaging (experimental)}
\usage{
ma(object, ...)

\S4method{ma}{a4aFitSAs}(object, stock, FUN, nsim = 1000)
}
\arguments{
\item{object}{an \code{a4aFits} object with the fits to be averaged across}

\item{...}{additional argument list that might never be used}

\item{stock}{a \code{stock} object with the original data used for fitting}

\item{FUN}{a \code{function} to compute the weights, which must return a named vector with weights. Note the weights will be normalized to sum 1 by \code{ma()}}

\item{nsim}{a \code{numeric} with the number of simulations to be drawn}
}
\value{
an \code{FLStock} object with iterations defined by \code{nsim}
}
\description{
Method to average across a set of models. This is still experimental. Use with care.
}
\examples{
data(ple4)
data(ple4.indices)
fmod <- ~ factor(age) + s(year, k=20)
qmod <- list(~ s(age, k = 4), ~ s(age, k = 4), ~ s(age, k = 3))
f1 <- sca(ple4, ple4.indices, fmodel=fmod, qmodel=qmod, fit = "assessment")
qmod <- list(~ s(age, k = 4)+year, ~ s(age, k = 4), ~ s(age, k = 3))
f2 <- sca(ple4, ple4.indices, fmodel=fmod, qmodel=qmod, fit = "assessment")
# AIC weighting
aicwt <- function(object){
 ICs <- -1 * sapply(object, AIC)
 exp( 0.5 * (ICs - max(ICs)))
}
stock.sim <- ma(a4aFitSAs(list(f1=f1, f2=f2)), ple4, aicwt, nsim = 100)
# equal weighting
eqwt <- function(object){
 v <- rep(1, length(object))
 names(v) <- names(object)
 v
}
stock.sim <- ma(a4aFitSAs(list(f1=f1, f2=f2)), ple4, eqwt, nsim = 100)
}

