% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addition-methods.R
\name{*}
\alias{*}
\alias{*,FLStock,a4aFitSA-method}
\alias{*,FLIndices,a4aFitSA-method}
\alias{*,FLStock,SCAPars-method}
\alias{*,FLIndices,SCAPars-method}
\title{* methods}
\usage{
\S4method{*}{FLStock,a4aFitSA}(e1, e2)

\S4method{*}{FLStock,SCAPars}(e1, e2)

\S4method{*}{FLIndices,a4aFitSA}(e1, e2)

\S4method{*}{FLIndices,SCAPars}(e1, e2)
}
\arguments{
\item{e1}{the original \code{FLStock} or \code{FLIndex} object}

\item{e2}{a \code{a4aFit} object from where the new \code{FLStock} or \code{FLIndex} slots will be extracted.}
}
\description{
Update \code{FLStock} and \code{FLIndex} objects with simulations from stock assessment fits.
}
