% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a4aGr-class.R
\docType{methods}
\name{niters,a4aGr-method}
\alias{niters,a4aGr-method}
\title{Number of iterations}
\usage{
\S4method{niters}{a4aGr}(object)
}
\arguments{
\item{object}{a \code{a4aGr} object}
}
\value{
an \code{numeric} object
}
\description{
Method to extract from \code{a4aGr} objects the number of iterations.
}
\examples{
mm <- matrix(NA, ncol=3, nrow=3)
diag(mm) <- c(50, 0.001,0.001)
mm[upper.tri(mm)] <- mm[lower.tri(mm)] <- c(0.1,0.01,0.00004)
vbObj <- a4aGr(grMod=~linf*(1-exp(-k*(t-t0))), grInvMod=~t0-1/k*log(1-len/linf), params=FLPar(linf=58.5, k=0.086, t0=0.001, units=c("cm","yr^-1","yr")), vcov=mm, distr="norm")
# Generate 100 sample sets
  vbObj <- mvrnorm(100,vbObj)
niters(vbObj)
}

