% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLModelSimMethods.R, R/utilities.R
\docType{methods}
\name{pars2dim}
\alias{pars2dim}
\alias{pars2dim-methods}
\alias{pars2dim,FLModelSim-method}
\alias{pars2dim,FLPar-method}
\title{Check that the second dimension in params is "iter"}
\usage{
pars2dim(object)

\S4method{pars2dim}{FLModelSim}(object)

\S4method{pars2dim}{FLPar}(object)
}
\arguments{
\item{object}{object of relevant class (see signature of method)}
}
\description{
Checks that the name of the second dimension in params is "iter". For internal use, not very interesting for users. It takes a \code{FLModelSim} object and returns a \code{logical}.
}
\section{pars2dim}{

Checks that the name of the second dimension in params is "iter". For internal use and not very interesting for users. It takes an \code{FLPar} object and returns a \code{logical}.
}

\examples{
pars2dim(FLModelSim())
#Example use of pars2dim:
pars2dim(FLPar())
pars2dim(FLPar(array(dim=c(1,1,1))))
}
