\name{pars2dim}
\alias{pars2dim}
\alias{pars2dim,FLModelSim-method}
\alias{pars2dim-methods}
\title{Check that the second dimension in params is "iter"}
\usage{
pars2dim(object)
}
\description{
Checks that the name of the second dimension in params is
"iter". For internal use, not very interesting for users.
It takes a \code{FLModelSim} object and returns a
\code{logical}
}
\examples{
pars2dim(FLModelSim())
}

