% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a4aFitresiduals-class.R
\docType{methods}
\name{plot of residuals}
\alias{plot of residuals}
\alias{plot,a4aFitResiduals,missing-method}
\title{Plot of standardized log residuals}
\usage{
\S4method{plot}{a4aFitResiduals,missing}(x, y = missing, ...)
}
\arguments{
\item{x}{an \code{a4aFitResiduals} object with the standardized residuals}

\item{y}{ignored}

\item{...}{additional argument list that might never be used}
}
\value{
a \code{plot} with stardardized log residuals
}
\description{
Method to produce scatterplots of standardized residuals
}
\examples{
data(ple4)
data(ple4.index)
obj <- sca(ple4, FLIndices(ple4.index))
flqs <- residuals(obj, ple4, FLIndices(idx=ple4.index))
plot(flqs)
}

