% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{methods}
\name{plot,a4aFit,FLStock-method}
\alias{plot,a4aFit,FLStock-method}
\title{plot of fitted catch numbers-at-age}
\usage{
\S4method{plot}{a4aFit,FLStock}(x, y, ...)
}
\arguments{
\item{x}{an \code{a4aFit} object with the fitted values}

\item{y}{an \code{FLStock} object with the observed values}

\item{...}{additional argument list that might never be used}
}
\value{
a \code{plot} with fitted and observed catch numbers-at-age
}
\description{
Method to plot fitted versus observed catch numbers-at-age.
}
\examples{
data(ple4)
data(ple4.index)
obj <- sca(ple4, FLIndices(ple4.index))
plot(obj, ple4)
}

