% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict-methods.R
\docType{methods}
\name{predict}
\alias{predict}
\alias{predict,SCAPars-method}
\alias{predict,a4aFitSA-method}
\title{predict methods for SCA}
\usage{
\S4method{predict}{a4aFitSA}(object)

\S4method{predict}{SCAPars}(object)
}
\description{
Predict methods for a4a stock assessment fits.
}
\examples{
data(ple4)
data(ple4.index)
fmodel <- ~factor(age) + factor(year)
qmodel <- list(~factor(age))
fit1 <-  a4aSCA(fmodel=fmodel, qmodel=qmodel, stock=ple4, indices=FLIndices(ple4.index))
flqs <- predict(fit1)
}

