% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{quant2mat}
\alias{quant2mat}
\title{coerces FLPar into list for call}
\usage{
quant2mat(x)
}
\value{
a list with each parametr as one element
}
\description{
coerces FLPar into list for call
}
\author{
EJ \email{ernesto.jardim@jrc.ec.europa.eu}
}
