% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redfish-data.R
\docType{data}
\name{rfLen}
\alias{rfLen}
\title{redfish length data}
\format{
An \code{FLStock}.
}
\source{
Daniel Howell
}
\usage{
data(rfLen)
}
\description{
Simulated length data for redfish.  Simulations were done using GADGET.
}
\author{
Ernesto Jardim
}
