% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rfTrawlTrd.idx}
\alias{rfTrawlTrd.idx}
\title{rfTrawlTrd.idx}
\format{an \code{FLIndex} object}
\usage{
rfTrawlTrd.idx
}
\description{
Trawl survey index for red fish, with a trend in catchability.
}
\author{
Daniel Howell
}
