\name{rngage}
\alias{rngage}
\alias{rngage,FLComp-method}
\alias{rngage-methods}
\alias{rngage<-}
\alias{rngage<-,FLComp,numeric-method}
\alias{rngage<--methods}
\title{age range extract and replacement}
\arguments{
  \item{object}{a \code{FLComp} object}

  \item{value}{a \code{vector} with max and min age range
  to replace the object info.}
}
\value{
  a \code{vector} object when extracting or a \code{FLComp}
  object when replacing
}
\description{
  Methods to extract from \code{FLComp} objects the age
  range or replace its value.
}

