% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCompMethods.R
\name{rngage}
\alias{rngage}
\alias{rngage-methods}
\alias{rngage,FLComp-method}
\alias{rngage<-}
\alias{rngage<--methods}
\alias{rngage<-,FLComp,numeric-method}
\title{age range extract and replacement}
\usage{
rngage(object, ...)
}
\arguments{
\item{object}{an \code{FLComp} object}

\item{value}{a \code{vector} with max and min age range to replace the object info}
}
\value{
a \code{vector} object when extracting, or an \code{FLComp} object when replacing
}
\description{
Methods to extract from \code{FLComp} objects the age range, or to replace its value.
}
\examples{
data(ple4)
data(ple4.index)
rngage(ple4)
rngage(ple4.index)
}
