\name{rnglen}
\alias{rnglen}
\alias{rnglen-methods}
\alias{rnglen,a4aGr-method}
\alias{rnglen<-}
\alias{rnglen<--methods}
\alias{rnglen<-,a4aGr,numeric-method}
\title{length range extract and replacement}
\arguments{
  \item{object}{a \code{a4aGr} object}

  \item{value}{a \code{vector} with max and min age range
  to replace the object info.}
}
\value{
  a \code{vector} object when extracting or a \code{a4aGr}
  object when replacing
}
\description{
  Methods to extract from \code{a4aGr} objects the length
  range or replace its value.
}

