% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLCompMethods.R
\name{rnglen}
\alias{rnglen}
\alias{rnglen,FLComp-method}
\alias{rnglen-methods}
\alias{rnglen<-}
\alias{rnglen<-,FLComp,numeric-method}
\alias{rnglen<--methods}
\title{length range extract and replacement}
\usage{
rnglen(object, ...)
}
\arguments{
\item{object}{a \code{FLComp} object}

\item{...}{Additional argument list that might not ever be used.}

\item{value}{a \code{vector} with max and min length range to replace the object info.}
}
\value{
a \code{vector} object when extracting or a \code{FLComp} object when replacing
}
\description{
Methods to extract from \code{FLComp} objects the length range or replace its value.
}

