\name{rnglen}
\alias{rnglen}
\alias{rnglen,a4aGr-method}
\alias{rnglen-methods}
\alias{rnglen<-}
\alias{rnglen<-,a4aGr,numeric-method}
\alias{rnglen<--methods}
\title{length range extract and replacement}
\usage{
rnglen(object, ...)
}
\arguments{
  \item{object}{a \code{a4aGr} object}

  \item{value}{a \code{vector} with max and min age range
  to replace the object info.}
}
\value{
a \code{vector} object when extracting or a \code{a4aGr}
object when replacing
}
\description{
Methods to extract from \code{a4aGr} objects the length
range or replace its value.
}

