% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCompMethods.R
\name{rnglen}
\alias{rnglen}
\alias{rnglen-methods}
\alias{rnglen,FLComp-method}
\alias{rnglen<-}
\alias{rnglen<--methods}
\alias{rnglen<-,FLComp,numeric-method}
\title{length range extract and replacement}
\usage{
rnglen(object, ...)
}
\arguments{
\item{object}{an \code{FLComp} object}

\item{value}{a \code{vector} with max and min length range to replace the object info}
}
\value{
a \code{vector} object when extracting, or an \code{FLComp} object when replacing
}
\description{
Methods to extract from \code{FLComp} objects the length range, or to replace its value.
}
\examples{
data(rfLen)
rnglen(rfTrawl.idx)
}
