\name{rngmbar}
\alias{rngmbar}
\alias{rngmbar-methods}
\alias{rngmbar,a4aM-method}
\alias{rngmbar<-}
\alias{rngmbar<--methods}
\alias{rngmbar<-,a4aM-method}
\title{M bar range extract and replacement}
\arguments{
  \item{object}{a \code{a4aM} object}

  \item{value}{a \code{vector} with max and min age range
  to replace the object info.}
}
\value{
  a \code{vector} object when extracting or a \code{a4aM}
  object when replacing
}
\description{
  Methods to extract from \code{a4aM} objects the M bar
  range or replace its value.
}

