\name{rngquant}
\alias{rngquant}
\alias{rngquant,FLComp-method}
\alias{rngquant-methods}
\alias{rngquant<-}
\alias{rngquant<-,FLComp,numeric-method}
\alias{rngquant<--methods}
\title{quant range extract and replacement}
\arguments{
  \item{object}{a \code{FLComp} object}

  \item{value}{a \code{vector} with max and min quant range
  to replace the object info.}
}
\value{
  a \code{vector} object when extracting or a \code{FLComp}
  object when replacing
}
\description{
  Methods to extract from \code{FLComp} objects the quant
  range or replace its value.
}

