% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCompMethods.R
\name{rngquant}
\alias{rngquant}
\alias{rngquant-methods}
\alias{rngquant,FLComp-method}
\alias{rngquant<-}
\alias{rngquant<--methods}
\alias{rngquant<-,FLComp,numeric-method}
\title{quant range extract and replacement}
\usage{
rngquant(object, ...)
}
\arguments{
\item{object}{an \code{FLComp} object}

\item{value}{a \code{vector} with max and min quant range to replace the object info}
}
\value{
a \code{vector} object when extracting, or an \code{FLComp} object when replacing
}
\description{
Methods to extract from \code{FLComp} objects the quant range, or to replace its value.
}
\examples{
data(ple4.index)
data(rfLen)
rngquant(ple4.index)
rngquant(rfTrawl.idx)
}
