% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLCompMethods.R
\name{rngquant}
\alias{rngquant}
\alias{rngquant,FLComp-method}
\alias{rngquant-methods}
\alias{rngquant<-}
\alias{rngquant<-,FLComp,numeric-method}
\alias{rngquant<--methods}
\title{quant range extract and replacement}
\usage{
rngquant(object, ...)
}
\arguments{
\item{object}{a \code{FLComp} object}

\item{...}{Additional argument list that might not ever be used.}

\item{value}{a \code{vector} with max and min quant range to replace the object info.}
}
\value{
a \code{vector} object when extracting or a \code{FLComp} object when replacing
}
\description{
Methods to extract from \code{FLComp} objects the quant range or replace its value.
}

