\name{rngyear}
\alias{rngyear}
\alias{rngyear,FLComp-method}
\alias{rngyear-methods}
\alias{rngyear<-}
\alias{rngyear<-,FLComp,numeric-method}
\alias{rngyear<--methods}
\title{year range extract and replacement}
\arguments{
  \item{object}{a \code{FLComp} object}

  \item{value}{a \code{vector} with max and min year range
  to replace the object info.}
}
\value{
  a \code{vector} object when extracting or a \code{FLComp}
  object when replacing
}
\description{
  Methods to extract from \code{FLComp} objects the year
  range or replace its value.
}

