% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fittingFunctions.R
\docType{methods}
\name{sca}
\alias{sca}
\alias{sca-methods}
\alias{sca,FLStock,FLIndices-method}
\alias{sca,FLStock,FLIndex-method}
\title{Statistical catch-at-age method}
\usage{
sca(stock, indices, ...)
}
\arguments{
\item{stock}{an \code{FLStock} object}

\item{indices}{an \code{FLIndices} object}

\item{fmodel}{a formula object depicting the model for log fishing mortality at age}

\item{qmodel}{a list of formula objects depicting the models for log survey catchability at age}

\item{srmodel}{a formula object depicting the model for log recruitment}

\item{fit}{character with type of fit: 'MP' or 'assessment'; the former does not require the hessian to be computed, while the latter does.}
}
\value{
an \code{a4aFit} or \code{a4aFitSA} object with the fit results.
}
\description{
User interface to the statistical catch-at-age method of the a4a stock assessment framework.
}
\examples{
data(ple4)
data(ple4.index)

# fit using the default submodels
fit1 <- sca(ple4, FLIndices(ple4.index))
plot(ple4 + fit1)

# see default submodels (set through automated procedure)
sca(ple4, FLIndices(ple4.index), fit='assessment')

# fishing mortality by age and year (separable)
fit2 <- sca(ple4, FLIndices(ple4.index), fmodel=~factor(age) + factor(year))
plot(ple4 + fit2)
wireframe(data~year*age, data=harvest(fit2), zlab="F")

# fit2 + catcability as a smoother by age without year trend
fit3 <- sca(ple4, FLIndices(ple4.index), fmodel=~factor(age) + factor(year), qmodel=list(~s(age, k=4)))
plot(ple4 + fit3)

# fit3 + srmodel as a smoother by year
fit4 <- sca(ple4, FLIndices(ple4.index), fmodel=~factor(age) + factor(year), qmodel=list(~s(age, k=4)), srmodel=~s(year, k=45))
plot(ple4 + fit4)

AIC(fit1, fit2, fit3, fit4)
BIC(fit1, fit2, fit3, fit4)
}
