% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCAMCMC-class.R
\docType{class}
\name{SCAMCMC}
\alias{SCAMCMC}
\alias{SCAMCMC-class}
\alias{SCAMCMC}
\alias{SCAMCMC-methods}
\alias{SCAMCMC,missing-method}
\alias{getADMBCallArgs}
\alias{mcmc}
\alias{mcmc-methods}
\alias{mcmc,SCAMCMC-method}
\alias{getN}
\alias{getADMBCallArgs}
\alias{getADMBCallArgs-methods}
\alias{SCAMCMC,missing-method}
\title{MCMC settings class}
\usage{
SCAMCMC(object, ...)

getADMBCallArgs(object, ...)

getN(object, ...)
}
\description{
Class definition (slots), constructors, accessors, replacement (when relevant) and common methods.
}
\section{Slot}{
 
\describe{

\item{\code{mcmc N}}{Run N MCMC iterations}
\item{\code{mcsave N}}{Save every N th MCMC iteration}
\item{\code{mcscale N}}{Rescale step size for first N iterations}
\item{\code{mcmult N}}{Rescale the covariance matrix}
\item{\code{mcrb N}}{Reduce high parameter correlations}
\item{\code{mcprobe X}}{Use a fat-tailed proposal distribution}
\item{\code{mcdiag}}{Use a diagonal covariance matrix}
\item{\code{mcnoscale}}{Do not scale the algorithm during}
\item{\code{mcu}}{Use a uniform distribution as proposal distribution}
\item{\code{hybrid}}{Use the hybrid method}
\item{\code{hynstep N}}{Mean number of steps for the leapfrog method}
\item{\code{hyeps X}}{The stepsize for the leapfrog method [X numeric and > 0]}

}
}

\section{Accessors}{

All slots in the class have accessor and replacement methods defined that
allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot.
A numeric vector can also be used when replacing FLQuant slots, and the
vector will be used to substitute the values in the slot, but not its other
attributes.
}

\section{Constructor}{

A construction method exists for this class that can take named arguments for
any of its slots. All slots are then created to match the requirements of the
class validity. If an unnamed \code{FLQuant} object is provided, this is used
for sizing, but not for populating any slot.
}

