% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-methods.R
\name{simulate}
\alias{simulate}
\alias{simulate,a4aFitSA-method}
\alias{simulate,SCAPars-method}
\alias{simulate,a4aStkParams-method}
\alias{simulate,submodels-method}
\alias{simulate,submodel-method}
\title{Simulation methods for SCA}
\usage{
simulate(object, nsim = 1, seed = NULL, ...)

\S4method{simulate}{a4aFitSA}(object, nsim = 1, seed = NULL, empirical = TRUE)

\S4method{simulate}{SCAPars}(object, nsim = 1, seed = NULL, empirical = TRUE)

\S4method{simulate}{a4aStkParams}(object, nsim = 1, seed = NULL, empirical = TRUE)

\S4method{simulate}{submodels}(object, nsim = 1, seed = NULL, empirical = TRUE)

\S4method{simulate}{submodel}(object, nsim = 1, seed = NULL, empirical = TRUE)
}
\arguments{
\item{object}{object of relevant class (see signature of method)}

\item{nsim}{number of iterations}

\item{seed}{\code{numeric} with random number seed}

\item{...}{additional argument list that might never be used}

\item{empirical}{logical, shall the empirical method in MASS be used}
}
\description{
Simulation methods for a4a stock assessment fits.
}
\examples{
data(ple4)
data(ple4.index)
fmodel <- ~factor(age) + factor(year)
qmodel <- list(~factor(age))
fit1 <-  sca(fmodel=fmodel, qmodel=qmodel, stock=ple4, indices=FLIndices(ple4.index))
fit1
summary(fit1)
stock.n(fit1)
}
