% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/a4aFitresiduals-class.R
\docType{methods}
\name{stdlogres}
\alias{stdlogres}
\alias{stdlogres,FLQuant,FLQuant-method}
\alias{stdlogres-methods}
\title{Standardized log residuals}
\usage{
stdlogres(obs, fit, ...)
}
\arguments{
\item{obs}{a \code{FLQuant} object with the observations}

\item{fit}{a \code{FLQuant} object with the fitted value}

\item{...}{Additional argument list that might not ever be used.}
}
\value{
a \code{FLQuant} with stardardized log residuals
}
\description{
Method o compute the standardized residuals in the log scale for the a4a stock assessment framework. Meaning indices and catch-at-age residuals.
}
\examples{
data(ple4)
data(ple4.index)
obj <- sca(ple4, FLIndices(ple4.index))
}

