% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodel-class.R
\docType{class}
\name{submodel}
\alias{submodel}
\alias{submodel-class}
\alias{submodel}
\alias{submodel-methods}
\alias{submodel,missing-method}
\alias{params,submodel-method}
\alias{sMod}
\alias{sMod-methods}
\alias{sMod,submodel-method}
\alias{vcov,submodel-method}
\alias{iter,submodel-method}
\alias{propagate,submodel-method}
\alias{formula,submodel-method}
\title{Submodel class}
\usage{
submodel(object, ...)

\S4method{submodel}{missing}(object, ...)

\S4method{params}{submodel}(object)

sMod(object, ...)

\S4method{sMod}{submodel}(object)

\S4method{vcov}{submodel}(object)

\S4method{iter}{submodel}(obj, it)

\S4method{propagate}{submodel}(object, iter, fill.iter = TRUE)

\S4method{formula}{submodel}(x)
}
\arguments{
\item{object}{object of relevant class (see signature of method)}

\item{...}{additional argument list that might never be used}

\item{obj}{the object to be subset}

\item{it}{iteration to be extracted}

\item{iter}{the number of iterations to create}

\item{fill.iter}{should the new iterations be filled with values (TRUE) or NAs (FALSE)}

\item{object}{the object to be extended}
}
\description{
Class definition (slots), constructors, accessors, replacement (when relevant) and common methods.
}
\section{Slot}{
 
\describe{

 \item{\code{Mod}}{\code{formula} describing the model}

 \item{\code{params}}{\code{FLPar} with model parameters}

 \item{\code{vcov}}{\code{array} with variance covariance paramaters related to the variance model}

 \item{\code{centering}}{\code{numeric} value used for centering the data}

 \item{\code{distr}}{a character with the parameters' statistical distribution; it must match a known distribution for R (\emph{e.g.} "norm" for gaussian) so that \code{rnorm} can be called}
}
}

\section{Accessors}{

All slots in the class have accessor and replacement methods defined that
allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot.
A numeric vector can also be used when replacing FLQuant slots, and the
vector will be used to substitute the values in the slot, but not its other
attributes.
}

\section{Constructor}{

A construction method exists for this class that can take named arguments for
any of its slots. All slots are then created to match the requirements of the
class validity. If an unnamed \code{FLQuant} object is provided, this is used
for sizing, but not for populating any slot.
}

