\name{vcov}
\alias{vcov}
\alias{vcov,SCAPars-method}
\alias{vcov,a4aFitSA-method}
\alias{vcov,submodel-method}
\alias{vcov,submodels-method}
\alias{vcov<-}
\alias{vcov<-,SCAPars,numeric-method}
\alias{vcov<-,a4aFitSA,numeric-method}
\alias{vcov<-,a4aStkParams,numeric-method}
\alias{vcov<-,submodel,numeric-method}
\alias{vcov<-,submodels,numeric-method}
\title{variance-covariance matrix}
\description{
  Methods to extract and replace the variance-covariance
  matrix
}

