% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLCompMethods.R
\name{vecage}
\alias{vecage}
\alias{vecage,FLComp-method}
\alias{vecage-methods}
\title{age vector extract}
\usage{
vecage(object, ...)
}
\arguments{
\item{object}{a \code{FLComp} object}

\item{...}{Additional argument list that might not ever be used.}
}
\value{
a \code{vector} object
}
\description{
method to extract from \code{FLComp} objects the vector of ages
}

