% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCompMethods.R
\name{vecage}
\alias{vecage}
\alias{vecage-methods}
\alias{vecage,FLComp-method}
\title{age vector extract}
\usage{
vecage(object, ...)
}
\arguments{
\item{object}{an \code{FLComp} object}
}
\value{
a \code{vector} object
}
\description{
Method to extract from \code{FLComp} objects the vector of ages.
}
\examples{
data(ple4)
data(ple4.index)
vecage(ple4)
vecage(ple4.index)
}
