% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCompMethods.R
\name{veclen}
\alias{veclen}
\alias{veclen-methods}
\alias{veclen,FLComp-method}
\title{length vector}
\usage{
veclen(object, ...)
}
\arguments{
\item{object}{an \code{FLComp} object}
}
\value{
a \code{vector} object
}
\description{
Method to extract from \code{FLComp} objects the vector of lengths.
}
\examples{
data(rfLen)
veclen(rfTrawl.idx)
}
