% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCompMethods.R
\docType{methods}
\name{vecquant}
\alias{vecquant}
\alias{vecquant,FLComp-method}
\alias{vecquant-methods}
\title{length vector}
\usage{
vecquant(object, ...)

\S4method{vecquant}{FLComp}(object)
}
\arguments{
\item{object}{object of relevant class (see signature of method)}

\item{...}{additional argument list that might never be used}
}
\value{
a \code{vector} object
}
\description{
method to extract from \code{FLComp} objects the vector of lengths.
}
\examples{
data(ple4.index)
data(rfLen)
vecquant(ple4.index)
vecquant(rfTrawl.idx)
}

