% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLCompMethods.R
\name{vecquant}
\alias{vecquant}
\alias{vecquant,FLComp-method}
\alias{vecquant-methods}
\title{length vector}
\usage{
vecquant(object, ...)
}
\arguments{
\item{object}{a \code{FLComp} object}
}
\value{
a \code{vector} object
}
\description{
method to extract from \code{FLComp} objects the vector of lengths.
}

