% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCompMethods.R
\docType{methods}
\name{vecyear}
\alias{vecyear}
\alias{vecyear,FLComp-method}
\alias{vecyear-methods}
\title{year vector extract}
\usage{
vecyear(object, ...)

\S4method{vecyear}{FLComp}(object)
}
\arguments{
\item{object}{object of relevant class (see signature of method)}

\item{...}{additional argument list that might never be used}
}
\value{
a \code{vector} object
}
\description{
Method to extract from \code{FLComp} objects the vector of years.
}
\examples{
data(ple4)
data(ple4.index)
vecyear(ple4)
vecyear(ple4.index)
}

